% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONSTANTS.R
\docType{data}
\name{API_CONSTANTS}
\alias{API_CONSTANTS}
\title{Load Package-wide Constants into Local Environment}
\format{
An object of class \code{character} of length 1.
}
\usage{
API_CONSTANTS
}
\description{
Load Package-wide Constants into Local Environment
}
\keyword{datasets}
