% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_db.R
\name{authenticate_db}
\alias{authenticate_db}
\title{Authenticate to Databrary.org.}
\usage{
authenticate_db(vb = FALSE)
}
\arguments{
\item{vb}{A boolean value.}
}
\value{
Status code if successful.
}
\description{
Authenticate to Databrary.org.
}
\examples{
authenticate_db()
}
