% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_db.R
\name{config_db}
\alias{config_db}
\title{Configure R environment for access to Databrary.org.}
\usage{
config_db(vb = FALSE)
}
\arguments{
\item{vb}{A boolean value.}
}
\description{
Configure R environment for access to Databrary.org.
}
\examples{
configure_db()
}
