% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_passwd.R
\name{config_passwd}
\alias{config_passwd}
\title{Store Databrary account credentials in secure system-level file.}
\usage{
config_passwd()
}
\description{
Store Databrary account credentials in secure system-level file.
}
\examples{
config_passwd()
}
