% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONSTANTS.R
\docType{data}
\name{DATABRARY_BASE_URL}
\alias{DATABRARY_BASE_URL}
\title{Load Package-wide Constants into Local Environment}
\format{
An object of class \code{character} of length 1.
}
\usage{
DATABRARY_BASE_URL
}
\description{
Load Package-wide Constants into Local Environment
}
\keyword{datasets}
