% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_csv.R
\name{download_csv}
\alias{download_csv}
\title{Downloads session spreadsheet as a CSV.}
\usage{
download_csv(vol_id = 1, file_name = "test.csv", target_dir = ".", vb = FALSE)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{file_name}{Name for the output file. Default is 'test.csv'.}

\item{target_dir}{Directory to save downloaded file. Default is '.'.}

\item{vb}{A boolean value.}
}
\value{
Name of downloaded file.
}
\description{
Downloads session spreadsheet as a CSV.
}
\examples{
download_csv() # Downloads "session" CSV for volume 1
}
