% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datavyu.R
\name{download_datavyu}
\alias{download_datavyu}
\title{Download a specific Datavyu (.opf) file.}
\usage{
download_datavyu(vol.id = 1, session.id = 9807, asset.id = 1,
  out.dir = "tmp", file.name = "test.opf", return.response = FALSE,
  vb = TRUE)
}
\arguments{
\item{vol.id}{Volume ID}

\item{session.id}{Slot/session number.}

\item{asset.id}{Asset number.}

\item{out.dir}{Directory to save output.}

\item{file.name}{Name for downloaded file.}

\item{return.response}{A Boolean value.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\description{
Download a specific Datavyu (.opf) file.
}
\examples{
download_datavyu()
}
