% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_party.R
\name{download_party}
\alias{download_party}
\title{Lists basic information about people on Databrary.}
\usage{
download_party(party = 6, to.df = FALSE, convert.JSON = TRUE,
  return.response = FALSE, vb = FALSE)
}
\arguments{
\item{party}{Party number to download.}

\item{to.df}{A Boolean value if TRUE returns a data frame.}

\item{convert.JSON}{A Boolean value if TRUE converts the JSON download}

\item{return.response}{A Boolean value if TRUE returns the GET response.}

\item{vb}{A Boolean value if TRUE returns verbose output.}
}
\value{
Status code if successful.
}
\description{
Lists basic information about people on Databrary.
}
\examples{
download_party()
}
