% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_party_avatar.R
\name{download_party_avatar}
\alias{download_party_avatar}
\title{Returns the avatar (image) for a given person.}
\usage{
download_party_avatar(party_id = 6, vb = FALSE)
}
\arguments{
\item{party_id}{Party number to retrieve information about.}

\item{vb}{A Boolean value if TRUE returns verbose output.}
}
\value{
Image file.
}
\description{
Returns the avatar (image) for a given person.
}
\examples{
download_party_avatar()
}
