% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_party_json.R
\name{download_party_json}
\alias{download_party_json}
\title{Lists basic information about people on Databrary.}
\usage{
download_party_json(party_id = 6, vb = FALSE)
}
\arguments{
\item{party_id}{Party number to retrieve information about.}

\item{vb}{A Boolean value if TRUE returns verbose output.}
}
\value{
JSON file with data.
}
\description{
Lists basic information about people on Databrary.
}
\examples{
download_party()
}
