% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_video.R
\name{download_vids_in_session}
\alias{download_vids_in_session}
\title{Download all videos in a session and save to a local directory.}
\usage{
download_vids_in_session(session_id = 9803, out_dir = ".", vb = FALSE)
}
\arguments{
\item{session_id}{Session number.}

\item{out_dir}{Directory to save output file.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\description{
Download all videos in a session and save to a local directory.
}
\examples{
download_vids_in_session()
}
