% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_video.R
\name{download_vids_in_volume}
\alias{download_vids_in_volume}
\title{Download all videos in a volume and save to a local directory.}
\usage{
download_vids_in_volume(vol_id = 31, out_dir = ".", vb = FALSE)
}
\arguments{
\item{out_dir}{Directory to save output file.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}

\item{session_id}{Volume number.}
}
\description{
Download all videos in a volume and save to a local directory.
}
\examples{
download_vids_in_volume()
}
