% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dv_code_defs.R
\name{extract_dv_code_defs}
\alias{extract_dv_code_defs}
\title{Extracts the code definitions from a Datavyu (.opf) file.}
\usage{
extract_dv_code_defs(in_dir = ".", dv_fn = "db", out_dir = in_dir,
  out_fn = paste0(out_dir, "/", "codes.csv"), auto_write_over = TRUE,
  code_regex = "^([a-zA-Z_]+[0-9]*[a-zA-Z_]*[0-9]*)",
  code_vals_regex = "\\\\)-([a-zA-Z\\\\/]+)\\\\|",
  code_type_regex = "([a-zA-Z]+)$", vb = FALSE)
}
\arguments{
\item{in_dir}{Input directory. Defaults to '.'}

\item{dv_fn}{Full Datavyu file name. Defaults to 'db'}

\item{out_dir}{Output directory. Defaults to in_dir}

\item{out_fn}{Full output file name.}

\item{auto_write_over}{A Boolean value. If TRUE, new output file overwrites old.}

\item{code_regex}{A string specifying the regular expression to extract a code.}

\item{code_vals_regex}{A string specifying the regular expression to extract a code value.}

\item{code_type_regex}{A string specifying the regular expression for the code type.}

\item{vb}{A boolean value. If TRUE, provides verbose output.}
}
\description{
Extracts the code definitions from a Datavyu (.opf) file.
}
\examples{
extract_dv_code_defs()
}
