% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_session_metadata}
\alias{extract_session_metadata}
\title{Extract Session Info}
\usage{
extract_session_metadata(volume_json)
}
\arguments{
\item{volume_json.}{A JSON blob returned by \code{get_volume_by_id}}
}
\description{
Extract Session Info
}
\examples{
\donttest{
\dontrun{
extract_session_metadata()
}
}
}
