% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_assets_in_session.R
\name{format_to_filetypes}
\alias{format_to_filetypes}
\title{Converts volume asset data frame to one with media types.}
\usage{
format_to_filetypes(vol_assets, vb = FALSE)
}
\description{
Converts volume asset data frame to one with media types.
}
