% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_affiliates.R
\name{get_affiliates}
\alias{get_affiliates}
\title{Lists affiliates (sponsored researchers) for a given Databrary party (institution or person).}
\usage{
get_affiliates(party_id = 6, vb = FALSE)
}
\arguments{
\item{party_id}{Target volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A tibble (data.frame) with the requested data.
}
\description{
Lists affiliates (sponsored researchers) for a given Databrary party (institution or person).
}
