% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_stats.R
\name{get_db_stats}
\alias{get_db_stats}
\title{Get Stats About Databrary}
\usage{
get_db_stats(type = "stats", vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{type}{Type of Databrary report to run "institutions", "people", "data"}

\item{vb}{Show verbose messages. Defaults to \code{options::opt("vb")}.}

\item{rq}{An \code{httr2} request object.}
}
\value{
A data frame with the requested data or NULL if there is
no new information.
}
\description{
Returns basic summary information about
the institutions, people, and video data hosted on Databrary.
}
\examples{
\dontrun{
get_db_stats()
get_db_stats("stats")
}
}
