% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_duration}
\alias{get_file_duration}
\title{Get Duration (In ms) Of A File.}
\usage{
get_file_duration(
  vol_id = 2,
  session_id = 9,
  asset_id = 2,
  types_w_durations = c(-600, -800),
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{vol_id}{Volume ID.}

\item{session_id}{Session ID containing the asset.}

\item{asset_id}{Asset number.}

\item{types_w_durations}{Asset types that have valid durations.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{rq}{An \code{httr2} request object. Default is NULL.}
}
\value{
Duration of a file in ms.
}
\description{
Get Duration (In ms) Of A File.
}
\examples{
\donttest{
get_file_duration() # default is a public video from volume 1
}

}
