% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_folder_file.R
\name{get_folder_file}
\alias{get_folder_file}
\title{Get Session File Data From A Databrary Volume}
\usage{
get_folder_file(
  vol_id = 1,
  folder_id = 9807,
  file_id = 1,
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{vol_id}{An integer indicating the volume identifier. Default is 1.}

\item{folder_id}{An integer indicating a valid folder identifier
linked to a volume. Default value is 9807, the materials folder for volume 1.}

\item{file_id}{An integer indicating the file identifier. Default is 1.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An httr2 request object.}
}
\value{
A JSON blob with the file data. If the user has previously logged
in to Databrary via \code{login_db()}, then files that have restricted access
can be downloaded, subject to the sharing release levels on those files.
}
\description{
Databrary volumes have folders where study or collection-wide files
can be stored and shared. \code{get_folder_file()} returns metadata about
specific files stored in a volume folder.
}
\examples{
\donttest{
\dontrun{
get_folder_file() # Data about file_id 1 from folder_id 9807 in Volume 1.
get_folder_file(vol_id = 2, folder_id = 9819, file_id = 16) # A PDF from Volume 2.
}
}
}
