% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_individual_sponsors.R
\name{get_individual_sponsors}
\alias{get_individual_sponsors}
\title{Lists a party's individual sponsors (people).}
\usage{
get_individual_sponsors(party_id = 4210, vb = FALSE)
}
\arguments{
\item{party_id}{Target volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A tibble (data.frame) with the requested data.
}
\description{
Lists a party's individual sponsors (people).
}
