% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_sessions.R
\name{get_info_from_session}
\alias{get_info_from_session}
\title{Helper function for list_volume_assets}
\usage{
get_info_from_session(volume_container, ignore_materials = FALSE)
}
\description{
Helper function for list_volume_assets
}
