% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_institution_by_id.R
\name{get_institution_by_id}
\alias{get_institution_by_id}
\title{Get institution metadata}
\usage{
get_institution_by_id(institution_id = 12, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{institution_id}{Institution identifier.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}
}
\value{
List of institution metadata or NULL when inaccessible.
}
\description{
Get institution metadata
}
