% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_permission_levels.R
\name{get_permission_levels}
\alias{get_permission_levels}
\title{Extracts Databrary permission level codes and labels.}
\usage{
get_permission_levels(vb = FALSE)
}
\arguments{
\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\description{
Extracts Databrary permission level codes and labels.
}
\examples{
get_permission_levels()
}
