% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_release_levels}
\alias{get_release_levels}
\title{Show Databrary Release Levels}
\usage{
get_release_levels(vb = options::opt("vb"))
}
\arguments{
\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}
}
\value{
A data frame with Databrary's release levels.
}
\description{
Show Databrary Release Levels
}
\examples{
\donttest{
get_release_levels()
}

}
