% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_as_df.R
\name{get_session_as_df}
\alias{get_session_as_df}
\title{Get Session Info From A Volume As A Data Frame.}
\usage{
get_session_as_df(vol_id = 1, vb = FALSE)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{vb}{A boolean value.}
}
\value{
A data frame with session information from a volume.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

This function has been superceded by \code{list_volume_sessions()}.
}
\examples{
\donttest{
get_session_as_df() # Return Session Info from Volume 1.
#' }

}
