% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_as_df.R
\name{get_session_csv}
\alias{get_session_csv}
\title{Get Session Info From A Volume and Return Data Frame}
\usage{
get_session_csv(vol_id = 1, vb = FALSE)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{vb}{A boolean value.}
}
\value{
List of assets.
}
\description{
Get Session Info From A Volume and Return Data Frame
}
\examples{
get_session_csv() # Downloads "session" CSV for volume 1
}
