% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sponsors.R
\name{get_sponsors}
\alias{get_sponsors}
\title{Lists sponsors for a given Databrary party (institution or person).}
\usage{
get_sponsors(party_id = 6, vb = FALSE)
}
\arguments{
\item{party_id}{Target volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A tibble (data.frame) with the requested data.
}
\description{
Lists sponsors for a given Databrary party (institution or person).
}
\examples{
get_sponsors() # Default is Rick Gilmore (party 6)
}
