% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_video_stats.R
\name{get_video_stats}
\alias{get_video_stats}
\title{Report Stats About Videos in a Volume.}
\usage{
get_video_stats(this_vol_id = 2, vb = FALSE)
}
\arguments{
\item{this_vol_id}{Target volume number.}

\item{vb}{A boolean value.}
}
\value{
A data frame with summary statistics about the videos in the volume.
}
\description{
#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
\code{get_video_stats()} has been superseded in favor of \code{summarize_videos_in_volume()}.
}
\examples{
\donttest{
\dontrun{
get_video_stats() # Retrieves summary statistics about videos in volume 2.
}
}
}
