% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_volume.R
\name{get_volume}
\alias{get_volume}
\title{Lists data for a given Databrary party (institution or person).}
\usage{
get_volume(vol_id = 1, search_str = "", vb = FALSE)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{search_str}{The specific data to retrieve for the volume}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A list (or data frame) with the requested data.
}
\description{
Lists data for a given Databrary party (institution or person).
}
\examples{
get_volume() # Default is volume 1
}
