% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HHMMSSmmm_to_ms}
\alias{HHMMSSmmm_to_ms}
\title{Convert Timestamp String To ms.}
\usage{
HHMMSSmmm_to_ms(HHMMSSmmm = "01:01:01:333")
}
\arguments{
\item{HHMMSSmmm}{a string in the format "HH:MM:SS:mmm"}
}
\description{
Convert Timestamp String To ms.
}
\examples{
HHMMSSmmm_to_ms() # 01:01:01:333 in ms
}
