% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_institution}
\alias{is_institution}
\title{Is This Party An Institution?}
\usage{
is_institution(party_id = 8, vb = FALSE)
}
\arguments{
\item{party_id}{Databrary party ID}

\item{vb}{A boolean value. If TRUE provides verbose output.}
}
\value{
TRUE or FALSEl.
}
\description{
Is This Party An Institution?
}
\examples{
is_institution() # Is party 8 (NYU) an institution.
}
