% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_person}
\alias{is_person}
\title{Is This Party A Person?}
\usage{
is_person(party_id = 7, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{party_id}{Databrary party ID}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{rq}{An \code{httr2} request object.}
}
\value{
TRUE if the party is a person, FALSE otherwise.
}
\description{
Is This Party A Person?
}
\examples{
\donttest{
is_person()
}

}
