% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_asset_formats.R
\name{list_asset_formats}
\alias{list_asset_formats}
\title{List Stored Assets (Files) By Type.}
\usage{
list_asset_formats(vb = options::opt("vb"))
}
\arguments{
\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}
}
\value{
A data frame with information about the data formats Databrary
supports.
}
\description{
List the data (file) formats supported by Databrary.
}
\examples{
\donttest{
list_asset_formats()
}
}
