% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_assets.R
\name{list_assets}
\alias{list_assets}
\title{Lists assets in a given Databrary volume and session (slot).}
\usage{
list_assets(slot = 9825, volume = 75, return.response = FALSE,
  convert.JSON = TRUE, vb = FALSE)
}
\arguments{
\item{slot}{Slot/session ID.}

\item{volume}{Selected volume number.}

\item{return.response}{A Boolean value. If TRUE returns results from GET.}

\item{convert.JSON}{A Boolean value. If TRUE converts JSON to data frame.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A plot of a volume's summary demographic characteristics.
}
\description{
Lists assets in a given Databrary volume and session (slot).
}
\examples{
list_assets()
}
