% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_assets_by_type.R
\name{list_assets_by_type}
\alias{list_assets_by_type}
\title{List Stored Assets (Files) By Type.}
\usage{
list_assets_by_type(vol_id = 2, type = "video", vb = FALSE)
}
\arguments{
\item{vol_id}{Volume ID}

\item{type}{Data file type, e.g. "video", "pdf", or "csv"}

\item{vb}{A boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with information about assets in a volume, sorted by
type.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated and may be removed from a future version of
the package. \code{list_volume_assets()} is preferred.
}
\examples{
\donttest{
list_assets_by_type() # Lists assets in volume 2.
}
}
