% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_authorized_investigators.R
\name{list_authorized_investigators}
\alias{list_authorized_investigators}
\title{List authorized investigators for an institution}
\usage{
list_authorized_investigators(
  institution_id = 12,
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{institution_id}{Institution identifier. Must be a positive integer. Default is 12.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An \code{httr2} request object. Defaults to \code{NULL}.}
}
\value{
Tibble of investigators; NULL if none.
}
\description{
Lists the authorized investigators at an institution.
}
