% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_containers_records_json.R
\name{list_containers_records_json}
\alias{list_containers_records_json}
\title{Lists containers  and record structure from Databrary volume.}
\usage{
list_containers_records_json(volume = 1, vb = FALSE)
}
\arguments{
\item{volume}{Databrary volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}

\item{convert.JSON}{Boolean.}
}
\value{
JSON file with containers and records from the specified volume.
}
\description{
Lists containers  and record structure from Databrary volume.
}
\examples{
list_containers_records_json()
}
