% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_folder_assets.R
\name{list_folder_assets}
\alias{list_folder_assets}
\title{List Assets Within a Databrary Folder.}
\usage{
list_folder_assets(
  folder_id = 1,
  vol_id = NULL,
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{folder_id}{Folder identifier scoped to the given volume.}

\item{vol_id}{Volume containing the folder. Required for Django API calls.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{rq}{An \code{httr2} request object. Defaults to \code{NULL}.}
}
\value{
A tibble with metadata for files contained in the folder, or
\code{NULL} when the folder has no accessible assets.
}
\description{
List Assets Within a Databrary Folder.
}
\examples{
\donttest{
\dontrun{
list_folder_assets(folder_id = 1, vol_id = 1)
}
}
}
