% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_institution_affiliates.R
\name{list_institution_affiliates}
\alias{list_institution_affiliates}
\title{List affiliates for an institution}
\usage{
list_institution_affiliates(
  institution_id = 12,
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{institution_id}{Institution identifier. Must be a positive integer. Default is 12.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An \code{httr2} request object. Defaults to \code{NULL}.}
}
\value{
Tibble of affiliates with roles and expiration dates.
}
\description{
List affiliates for an institution
}
