% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_party.R
\name{list_party}
\alias{list_party}
\title{Lists data for a given Databrary party (institution or person).}
\usage{
list_party(party_id = 8, component = "all", vb = FALSE)
}
\arguments{
\item{party_id}{Target volume number.}

\item{component}{The specific data to retrieve for the party, c("all", "children", "parents")}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A list (or data frame) with the requested data.
}
\description{
Lists data for a given Databrary party (institution or person).
}
\examples{
list_party() # Default is New York University (party 8)
}
