% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_party_affiliates.R
\name{list_party_affiliates}
\alias{list_party_affiliates}
\title{List Affiliates For A Party}
\usage{
list_party_affiliates(party_id = 6, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{party_id}{Target party ID.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{rq}{An \code{httr2} request object. Defaults to NULL.}
}
\value{
A data frame with information about a party's affiliates.
}
\description{
List Affiliates For A Party
}
\examples{
\donttest{
list_party_affiliates() # Default is Rick Gilmore (party 6)
}
}
