% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_party_volumes.R
\name{list_party_volumes}
\alias{list_party_volumes}
\title{List Volumes A Party Has Access To}
\usage{
list_party_volumes(party_id = 6, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{party_id}{Target party ID.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}

\item{rq}{An \code{httr2}-style request object. If NULL, then a new request will
be generated using \code{make_default_request()}.}
}
\value{
A data frame with information about a party's sponsors.
}
\description{
List Volumes A Party Has Access To
}
\examples{
\donttest{
\dontrun{
list_party_volumes() # Default is Rick Gilmore (party 6)
}
}
}
