% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_session_activity.R
\name{list_session_activity}
\alias{list_session_activity}
\title{List Activity History in Databrary Session.}
\usage{
list_session_activity(
  vol_id = 1892,
  session_id = 76113,
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{vol_id}{Volume identifier (required by the Django API). Must be a positive integer.}

\item{session_id}{Session identifier. Must be a positive integer.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An \code{httr2} request object. Defaults to \code{NULL}. When \code{NULL}, a
default request is generated, but this will only permit public information
to be returned.}
}
\value{
A tibble with the activity history for a session, or \code{NULL} when
no data is available.
}
\description{
For an accessible session, returns the logged history events associated with
the session. Requires authenticated access with sufficient permissions.
}
\examples{
\\donttest{
\\dontrun{
list_session_activity(vol_id = 1892, session_id = 76113)
}
}
}
