% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_session_assets_2.R
\name{list_session_assets_2}
\alias{list_session_assets_2}
\title{List Assets in a Session from a Databrary volume.}
\usage{
list_session_assets_2(session_id = 9807, vb = FALSE, rq = NULL)
}
\arguments{
\item{session_id}{An integer. A Databrary session number. Default is 9807,
the "materials" folder from Databrary volume 1.}

\item{vb}{A logical value. Show verbose feedback. Default is FALSE.}

\item{rq}{An \code{httr2} request object. If NULL, a default request is generated
from databraryr::make_default_request().}
}
\value{
A data frame with information about all assets in a volume.
}
\description{
#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
\code{list_session_assets_2()} may eventually replace  \code{list_session_assets()} as
the new function does not require a volume id.
}
\examples{
\donttest{
\dontrun{
list_session_assets_2() # Session 9807 in volume 1
}
}
}
