% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_assets_in_session.R
\name{list_specified_assets_in_session}
\alias{list_specified_assets_in_session}
\title{Lists assets in a given Databrary session (slot).}
\usage{
list_specified_assets_in_session(
  session_id = 9807,
  media_type = "MPEG-4 video",
  vb = FALSE
)
}
\arguments{
\item{session_id}{Slot/session ID.}

\item{media_type}{A string indicating what type of file.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with the assets in the selected volume and session.
}
\description{
Lists assets in a given Databrary session (slot).
}
