% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_user_affiliates.R
\name{list_user_affiliates}
\alias{list_user_affiliates}
\title{List affiliates for a user}
\usage{
list_user_affiliates(user_id = 6, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{user_id}{User identifier. Must be an integer. Default is 6.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An \code{httr2} request object. Defaults to \code{NULL}.}
}
\value{
Tibble of affiliates for the user.
}
\description{
List affiliates for a user
}
