% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_user_history.R
\name{list_user_history}
\alias{list_user_history}
\title{List Account Activity For A Databrary User.}
\usage{
list_user_history(user_id = 22582, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{user_id}{Target user identifier.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{rq}{An \code{httr2} request object. Defaults to \code{NULL}.}
}
\value{
A tibble containing authentication and activity events for the
selected user, or \code{NULL} when no entries are available.
}
\description{
Retrieve the OAuth and login activity history for a specific
user. Access is restricted to administrators and authorized investigators
with sufficient privileges.
}
\examples{
\donttest{
\dontrun{
list_user_history(user_id = 22582)
}
}
}
