% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_user_sponsors.R
\name{list_user_sponsors}
\alias{list_user_sponsors}
\title{List sponsorships for a user}
\usage{
list_user_sponsors(user_id = 6, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{user_id}{User identifier.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}
}
\value{
Tibble of sponsors for the user.
}
\description{
List sponsorships for a user
}
