% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_user_volumes.R
\name{list_user_volumes}
\alias{list_user_volumes}
\title{List Volumes Associated With A User}
\usage{
list_user_volumes(user_id = 6, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{user_id}{User identifier. Must be a positive integer. Default is 6.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An \code{httr2} request object. Default is NULL.}
}
\value{
Tibble of volumes the user owns or collaborates on.
}
\description{
List Volumes Associated With A User
}
