% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume.R
\name{list_volume}
\alias{list_volume}
\title{List Info for a Databrary Volume.}
\usage{
list_volume(vol_id = 1, search_str = "", vb = FALSE)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{search_str}{The specific data to retrieve for the volume}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with information about a volume.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{list_volume()} has been deprecated in lieu of other, more specific,
volume-related functions. It will be dropped from a future release.
}
\examples{
\donttest{
\dontrun{
list_volume() # Default is volume 1
}
}
}
