% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_activity.R
\name{list_volume_activity}
\alias{list_volume_activity}
\title{List Activity In A Databrary Volume}
\usage{
list_volume_activity(vol_id = NULL, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{vol_id}{Selected volume number. Must be a positive integer. Default is NULL.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An \code{httr2} request object. Defaults to NULL.}
}
\value{
A list with the activity history on a volume.
}
\description{
If a user has access to a volume, this command lists the modification
history of the volume.
}
\examples{
\donttest{
\dontrun{
# The following will only return output if the user has *write* privileges
# on the volume.

list_volume_activity(vol_id)
}
}
}
