% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_folders.R
\name{list_volume_folders}
\alias{list_volume_folders}
\title{List Folders in a Databrary Volume.}
\usage{
list_volume_folders(vol_id = 1, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{vol_id}{Target volume number. Must be a positive integer. Default is 1.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An \code{httr2} request object. Defaults to \code{NULL}.}
}
\value{
A tibble with metadata about folders in the selected volume, or
\code{NULL} when no folders are available.
}
\description{
List Folders in a Databrary Volume.
}
\examples{
\donttest{
\dontrun{
list_volume_folders() # Folders in volume 1
}
}
}
