% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_links.R
\name{list_volume_links}
\alias{list_volume_links}
\title{Retrieves URL Links From A Databrary Volume.}
\usage{
list_volume_links(vol_id = 1, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{vol_id}{Target volume number. Must be a positive integer. Default is 1.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An \code{httr2} request object.}
}
\value{
A data frame with the requested data.
}
\description{
Retrieves URL Links From A Databrary Volume.
}
\examples{
\donttest{
\dontrun{
list_volume_links() # Links from volume 1
}
}
}
