% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_metadata.R
\name{list_volume_metadata}
\alias{list_volume_metadata}
\title{List volume metadata.}
\usage{
list_volume_metadata(volume = 2, write.header = FALSE,
  data.frame = TRUE, vb = FALSE)
}
\arguments{
\item{volume}{Selected volume number.}

\item{write.header}{A Boolean value. If TRUE writes a comma-separated header.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with information about the volume.
}
\description{
List volume metadata.
}
\examples{
list_volume_metadata()
}
