% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_owners.R
\name{list_volume_owners}
\alias{list_volume_owners}
\title{List volume owners.}
\usage{
list_volume_owners(volume = 1)
}
\arguments{
\item{volume}{Selected volume number. Default is 2.}
}
\value{
A data framw with information about the volume owner(s).
}
\description{
List volume owners.
}
\examples{
list_volume_owners()
}
