% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_sessions.R
\name{list_volume_sessions}
\alias{list_volume_sessions}
\title{List Sessions in Databrary Volume.}
\usage{
list_volume_sessions(
  vol_id = 1,
  include_vol_data = FALSE,
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{vol_id}{Target volume number. Must be a positive integer. Default is 1.}

\item{include_vol_data}{A Boolean value. Include volume-level metadata
or not. Default is FALSE.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An \code{httr2} request object. If NULL (the default)
a request will be generated, but this will only permit public information
to be returned.}
}
\value{
A data frame with information about all assets in a volume.
}
\description{
List Sessions in Databrary Volume.
}
\examples{
\donttest{
\dontrun{
list_volume_sessions() # Sessions in Volume 1
}
}
}
