% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_tags.R
\name{list_volume_tags}
\alias{list_volume_tags}
\title{Lists Keywords And Tags For A Volume.}
\usage{
list_volume_tags(vol_id = 1, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{vol_id}{Target volume number. Must be a positive integer. Default is 1.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An \code{httr2} request object. Default is NULL.}
}
\value{
A data frame with the requested data.
}
\description{
Lists Keywords And Tags For A Volume.
}
\examples{
\dontrun{
list_volume_tags()
}
}
