% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout_db.R
\name{logout_db}
\alias{logout_db}
\title{Log Out of Databrary.org.}
\usage{
logout_db(vb = options::opt("vb"))
}
\arguments{
\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{rq}{An \code{httr2} request object. Defaults to NULL.}
}
\value{
TRUE if logging out succeeds, FALSE otherwise.
}
\description{
Log Out of Databrary.org.
}
\examples{
\donttest{
logout_db()
}
}
