% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{options}
\alias{options}
\title{databraryr Options}
\description{
Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.
}
\section{Checking Option Values}{

Option values specific to \code{databraryr} can be
accessed by passing the package name to \code{env}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{options::opts(env = "databraryr")

options::opt(x, default, env = "databraryr")
}\if{html}{\out{</div>}}
}

\section{Options}{

\describe{
\item{vb}{\describe{
Show verbose messages.\item{default: }{\preformatted{TRUE}}
\item{option: }{databraryr.vb}
\item{envvar: }{R_DATABRARYR_VB (evaluated if possible, raw string otherwise)}
}}

\item{quiet}{\describe{
Be quiet in messaging\item{default: }{\preformatted{TRUE}}
\item{option: }{databraryr.quiet}
\item{envvar: }{R_DATABRARYR_QUIET (evaluated if possible, raw string otherwise)}
}}

}
}

\seealso{
options getOption Sys.setenv Sys.getenv
}
