% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_data_as_df.R
\name{read_csv_data_as_df}
\alias{read_csv_data_as_df}
\title{Download a stored CSV data file as a data.frame.}
\usage{
read_csv_data_as_df(session_id = 9807, asset_id = 116888, vb = FALSE)
}
\arguments{
\item{session_id}{Session ID within some volume.}

\item{asset_id}{Asset ID within the session.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\description{
Download a stored CSV data file as a data.frame.
}
\examples{
read_csv_data_as_df() # Loads CSV depicting Databrary growth from volume 1
}
