% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_keywords.R
\name{search_for_keywords}
\alias{search_for_keywords}
\title{Search For Keywords in Databrary Volumes.}
\usage{
search_for_keywords(
  search_string = "locomotion",
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{search_string}{String to search.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{rq}{An \code{httr2} request object. Default is NULL.}
}
\value{
A list with the volumes that contain the keyword.
}
\description{
Search For Keywords in Databrary Volumes.
}
\examples{
\dontrun{
search_for_keywords() # searches for volumes with "locomotion" as a keyword.
search_for_keywords()

# searches for volumes with "adult" as a keyword.
search_for_keywords("adult") 
}
}
