% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_demo_part_w_video.R
\name{summ_demo_part_w_vid}
\alias{summ_demo_part_w_vid}
\title{Summarize demographic data for participants in sessions with videos}
\usage{
summ_demo_part_w_vid(vol.id = 4, vb = FALSE)
}
\arguments{
\item{vol.id}{Selected volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
Data frame with the vol.id, number of videos, and total hours.
}
\description{
Summarize demographic data for participants in sessions with videos
}
\examples{
summarize_videos_in_volume()
}
