% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_demo_part_w_video.R
\name{summarize_demo_part_w_video}
\alias{summarize_demo_part_w_video}
\title{Summarize demographic data for participants in sessions with videos}
\usage{
summarize_demo_part_w_video(vol_id = 4, vb = FALSE)
}
\arguments{
\item{vol_id}{Selected volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
Data frame with the vol_id, number of videos, and total hours.
}
\description{
Summarize demographic data for participants in sessions with videos
}
\examples{
summarize_demo_part_w_vid()
}
