% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_demog.R
\name{summarize_demog}
\alias{summarize_demog}
\title{Plot summary of volume's participant characteristics.}
\usage{
summarize_demog(volume = 4, return.df = FALSE, vb = FALSE)
}
\arguments{
\item{volume}{Selected volume number.}

\item{return.df}{A boolean value.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
Status code if successful.
}
\description{
Plot summary of volume's participant characteristics.
}
\examples{
summarize_demog()
}
