% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unassign_record_from_file.R
\name{unassign_record_from_file}
\alias{unassign_record_from_file}
\title{Unassign Record from Session File}
\usage{
unassign_record_from_file(
  vol_id = 1,
  session_id,
  file_id,
  record_id,
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{vol_id}{Target volume number. Must be a positive integer.}

\item{session_id}{Numeric session identifier. Must be a positive integer.}

\item{file_id}{Numeric file identifier. Must be a positive integer.}

\item{record_id}{Numeric record identifier. Must be a positive integer.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{rq}{An \code{httr2} request object. Defaults to \code{NULL}.}
}
\value{
\code{TRUE} if the record was successfully unassigned, \code{FALSE}
otherwise.
}
\description{
Remove the association between a record and a session file.
If the record is not assigned to the file, this operation will fail.
}
\examples{
\donttest{
\dontrun{
# Unassign a record from a file
unassign_record_from_file(
  vol_id = 1,
  session_id = 10,
  file_id = 20,
  record_id = 123
)
}
}
}
