% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_video.R
\name{validate_segment_id}
\alias{validate_segment_id}
\title{Download all videos in a session and save to a local directory.}
\usage{
validate_segment_id(
  segment_id,
  video_duration_ms,
  segment_regex = "([0-9]+),([0-9]+)",
  vb = FALSE
)
}
\description{
Download all videos in a session and save to a local directory.
}
