% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_constants.R
\name{assign_constants}
\alias{assign_constants}
\title{Download Databrary Constants From API.}
\usage{
assign_constants(vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{vb}{A Boolean value. If TRUE provides verbose output.}

\item{rq}{An \code{httr2} request object. Defaults to NULL.}
}
\value{
A data frame with the constants.
}
\description{
Download Databrary Constants From API.
}
\examples{
\donttest{
assign_constants()
}
}
