% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_constants.R
\name{assign_constants}
\alias{assign_constants}
\title{Downloads Databrary contants from API.}
\usage{
assign_constants(vb = FALSE, convert.JSON = TRUE)
}
\arguments{
\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with the constants.
}
\description{
Downloads Databrary contants from API.
}
\examples{
assign_constants()
}
