% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_databrary_ssl_certs.R
\name{check_ssl_certs}
\alias{check_ssl_certs}
\title{Checks the SSL certificates for nyu.databary.org.}
\usage{
check_ssl_certs(host = "nyu.databrary.org")
}
\arguments{
\item{url}{Target URL. Defaults to 'nyu.databrary.org'.}
}
\description{
Checks the SSL certificates for nyu.databary.org.
}
\examples{
check_ssl_certs()
}
