% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_session.R
\name{create_session}
\alias{create_session}
\title{Create New Session on A Databrary Volume}
\usage{
create_session(vol_id, rq = NULL, lrsp = NULL)
}
\arguments{
\item{vol_id}{Volume ID.}

\item{rq}{An httr2 request object.}

\item{lrsp}{An httr2 response object returned from \code{login_db()}}
}
\value{
A JSON blob with the new session's data if the user has previously logged
in to Databrary via \code{login_db()}.
}
\description{
Create New Session on A Databrary Volume
}
\examples{
\donttest{
create_session() # Default is Volume 1
}
}
