% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databraryr-package.R
\docType{package}
\name{databraryr-package}
\alias{databraryr}
\alias{databraryr-package}
\title{databraryr: Interact with the 'Databrary.org' API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

'Databrary.org' is a restricted access repository for research data, especially video and audio. This package provides commands to interact with the data stored on 'Databrary.org'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://databrary.github.io/databraryr/}
  \item \url{https://github.com/databrary/databraryr}
  \item Report bugs at \url{https://github.com/databrary/databraryr/issues}
}

}
\author{
\strong{Maintainer}: Rick O. Gilmore \email{rog1@psu.edu} [copyright holder]

Authors:
\itemize{
  \item Jeffrey Spies \email{cran@jeffspies.com}
}

Other contributors:
\itemize{
  \item National Science Foundation OAC-2032713 [funder]
  \item National Institutes of Health R01HD094830 [funder]
}

}
\keyword{internal}
