% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databraryr-package.R
\docType{package}
\name{databraryr-package}
\alias{databraryr}
\alias{databraryr-package}
\title{databraryr: Interacts With the Databrary.org API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides commands to interact with the Databrary.org API. There are some public assets on Databrary. But for most interesting use cases, a user will need to register \url{https://databrary.org/register} and receive authorization from an institution. Once authorized, users can download data and metadata from Databrary and reuse it subject to the terms of the Databrary Access Agreement and the site Terms and Conditions of Use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/databrary/databraryr}
  \item Report bugs at \url{https://github.com/databrary/databraryr/issues}
}

}
\author{
\strong{Maintainer}: Rick O. Gilmore \email{rog1@psu.edu} [copyright holder]

Authors:
\itemize{
  \item Jeffrey Spies \email{cran@jeffspies.com}
}

Other contributors:
\itemize{
  \item National Science Foundation OAC-2032713 [funder]
  \item National Institutes of Health R01HD094830 [funder]
}

}
\keyword{internal}
