% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databraryr-package.R
\docType{package}
\name{databraryr-package}
\alias{databraryr}
\alias{databraryr-package}
\title{databraryr: Interacts With the Databrary.org API}
\description{
This package provides commands to interact with the Databrary.org API. There are some public assets on Databrary. But for most interesting use cases, a user will need to register \url{http://databrary.org/register} and receive authorization from an institution. Once authorized, users can download data and metadata from Databrary.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/databrary/databraryr}
  \item Report bugs at \url{http://github.com/databrary/databraryr/issues}
}

}
\author{
\strong{Maintainer}: Rick O. Gilmore \email{rog1@psu.edu}

}
\keyword{internal}
