% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_asset.R
\name{download_asset}
\alias{download_asset}
\title{Download a specific video file.}
\usage{
download_asset(asset = 11643, slot = 9825, file.name = "test.mp4",
  return.response = FALSE, vb = FALSE)
}
\arguments{
\item{asset}{Asset number.}

\item{slot}{Slot/session number.}

\item{file.name}{Name for downloaded file.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}

\item{return.respose}{A Boolean value.}
}
\description{
Download a specific video file.
}
\examples{
download_asset()
}
