% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_asset.R
\name{download_asset}
\alias{download_asset}
\title{Download Asset From Databrary.}
\usage{
download_asset(
  asset_id = 1,
  session_id = 9807,
  file_name = "test.mp4",
  target_dir = tempdir(),
  return_response = FALSE,
  vb = FALSE
)
}
\arguments{
\item{asset_id}{Asset id for target file.}

\item{session_id}{Slot/session number where target file is stored.}

\item{file_name}{Name for downloaded file.}

\item{target_dir}{Directory to save the downloaded file.
Default is a temporary directory given by a call to \code{tempdir()}.}

\item{return_response}{A Boolean value.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
Full file name to the asset.
}
\description{
Download Asset From Databrary.
}
\examples{
download_asset() # Download's 'numbers' file from volume 1.
download_asset(asset_id = 11643, session_id = 9825, file_name = "rdk.mp4") 
# Downloads a display with a random dot kinematogram (RDK).
}
