% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_asset.R
\name{download_asset}
\alias{download_asset}
\title{Download a specific video, audio, or other file.}
\usage{
download_asset(
  asset_id = 11643,
  session_id = 9825,
  file_name = "test.mp4",
  return_response = FALSE,
  vb = FALSE
)
}
\arguments{
\item{asset_id}{Asset number.}

\item{session_id}{Slot/session number.}

\item{file_name}{Name for downloaded file.}

\item{return_response}{A Boolean value.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\description{
Download a specific video, audio, or other file.
}
\examples{
download_asset()
}
