% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_containers_records.R
\name{download_containers_records}
\alias{download_containers_records}
\title{Downloads container and record structure from Databrary volume.}
\usage{
download_containers_records(vol.id = 2, convert.JSON = TRUE,
  vb = FALSE)
}
\arguments{
\item{vol.id}{Databrary volume number.}

\item{convert.JSON}{Boolean.}

\item{vb}{A boolean value. If TRUE provides verbose output.}
}
\value{
List of containers and records from the specified volume.
}
\description{
Downloads container and record structure from Databrary volume.
}
\examples{
downloade_containers_records()
}
