% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datavyu.R
\name{download_datavyu}
\alias{download_datavyu}
\title{Download a specific Datavyu (.opf) file.}
\usage{
download_datavyu(vol_id = 1, session_id = 9807, asset_id = 117035,
  out_dir = "tmp", file_name = "test.opf", return_response = FALSE,
  vb = TRUE)
}
\arguments{
\item{vol_id}{Volume ID}

\item{session_id}{Slot/session number.}

\item{asset_id}{Asset number.}

\item{out_dir}{Directory to save output.}

\item{file_name}{Name for downloaded file.}

\item{return_response}{A Boolean value.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\description{
Download a specific Datavyu (.opf) file.
}
\examples{
download_datavyu()
}
