% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_party.R
\name{download_party}
\alias{download_party}
\title{Lists basic information about people on Databrary.}
\usage{
download_party(party_id = 6, convert_JSON = TRUE, vb = FALSE)
}
\arguments{
\item{party_id}{Party number to retrieve information about.}

\item{convert_JSON}{A Boolean value if TRUE converts the JSON download}

\item{vb}{A Boolean value if TRUE returns verbose output.}
}
\value{
Status code if successful.
}
\description{
#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
\code{download_party()} has been superseded in favor of \code{get_party_as_df()}.
}
\examples{
download_party()
}
