% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_party.R
\name{download_party}
\alias{download_party}
\title{Lists basic information about people on Databrary.}
\usage{
download_party(party.id = 6, convert.JSON = TRUE, vb = FALSE)
}
\arguments{
\item{party.id}{Party number to retrieve information about.}

\item{convert.JSON}{A Boolean value if TRUE converts the JSON download}

\item{vb}{A Boolean value if TRUE returns verbose output.}
}
\value{
Status code if successful.
}
\description{
Lists basic information about people on Databrary.
}
\examples{
download_party()
}
