% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_party.R
\name{download_party}
\alias{download_party}
\title{Download Information About a Party on Databrary.}
\usage{
download_party(party_id = 6, convert_JSON = TRUE, vb = FALSE)
}
\arguments{
\item{party_id}{Party number to retrieve information about.}

\item{convert_JSON}{A Boolean value if TRUE converts the JSON download}

\item{vb}{A Boolean value if TRUE returns verbose output.}
}
\value{
A data frame with information about the party.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has been deprecated and may be removed in a future release.
See \code{get_party_by_id()} for similar, and more general functionality.
}
\examples{
\donttest{
download_party()
#' }

}
