% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_party_avatar.R
\name{download_party_avatar}
\alias{download_party_avatar}
\title{Returns the avatar (image) for a given person.}
\usage{
download_party_avatar(party_id = 6, show_person_info = TRUE, vb = FALSE)
}
\arguments{
\item{party_id}{Party number to retrieve information about. Default is 6
(Rick Gilmore).}

\item{show_person_info}{Show the person's name and affiliation in the output.
Default is TRUE.}

\item{vb}{A Boolean value if TRUE returns verbose output. Default is TRUE.}
}
\value{
An image file.
}
\description{
Returns the avatar (image) for a given person.
}
\examples{
download_party_avatar() # Show Rick Gilmore's (party 6) avatar.
}
