% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_session_csv.R
\name{download_session_csv}
\alias{download_session_csv}
\title{Downloads session spreadsheet as a CSV.}
\usage{
download_session_csv(volume = 1, to.df = TRUE, return.response = FALSE,
  vb = FALSE)
}
\arguments{
\item{volume}{Target volume number.}

\item{to.df}{A boolean value.}

\item{return.response}{A boolean value.}

\item{vb}{A boolean value.}
}
\value{
List of assets.
}
\description{
Downloads session spreadsheet as a CSV.
}
\examples{
download_csv()
}
