% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_session_csv.R
\name{download_session_csv}
\alias{download_session_csv}
\title{Download Session Spreadsheet As CSV}
\usage{
download_session_csv(
  vol_id = 1,
  file_name = "test.csv",
  target_dir = tempdir(),
  as_df = FALSE,
  vb = FALSE,
  rq = NULL
)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{file_name}{Name for the output file. Default is 'test.csv'.}

\item{target_dir}{Directory to save downloaded file. Default is tempdir().}

\item{as_df}{A Boolean value. Default is FALSE.}

\item{vb}{A Boolean value. Default is FALSE.}

\item{req}{An \code{httr2} request object.}
}
\value{
A character string that is the name of the downloaded file or a data frame if \code{as_df} is TRUE.
}
\description{
Databrary generates a CSV-formated spreadsheet that summarizes
information about individual sessions. This command downloads that CSV file
as a temporary file or with a name specified by the user.
}
\examples{
\donttest{
\dontrun{
download_session_csv() # Downloads "session" CSV for volume 1
}
}
}
