% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_session_csv.R
\name{download_session_csv}
\alias{download_session_csv}
\title{Downloads session spreadsheet as a CSV.}
\usage{
download_session_csv(vol_id = 1, to_df = TRUE,
  return_response = FALSE, vb = FALSE)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{to_df}{A boolean value.}

\item{return_response}{A boolean value.}

\item{vb}{A boolean value.}
}
\value{
List of assets.
}
\description{
Downloads session spreadsheet as a CSV.
}
\examples{
download_csv()
}
