% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_session_zip.R
\name{download_session_zip}
\alias{download_session_zip}
\title{Download a zip archive with all data from a specified volume and session}
\usage{
download_session_zip(
  vol_id = 31,
  session_id = 9803,
  out_dir = tempdir(),
  file_name = "test.zip",
  vb = FALSE
)
}
\arguments{
\item{vol_id}{Volume number.}

\item{session_id}{Slot/session number.}

\item{out_dir}{Directory to save output file.}

\item{file_name}{Name for downloaded file, default is 'test.zip'.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\description{
Download a zip archive with all data from a specified volume and session
}
\examples{
download_session_zip()
}
