% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_session_zip.R
\name{download_session_zip}
\alias{download_session_zip}
\title{Download Zip Archive From Databrary Session.}
\usage{
download_session_zip(
  vol_id = 31,
  session_id = 9803,
  out_dir = tempdir(),
  file_name = "test.zip",
  vb = FALSE
)
}
\arguments{
\item{vol_id}{Volume number.}

\item{session_id}{Slot/session number.}

\item{out_dir}{Directory to save output file.}

\item{file_name}{Name for downloaded file, default is 'test.zip'.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
Full filename of the downloaded file.
}
\description{
Download Zip Archive From Databrary Session.
}
\examples{
\donttest{
\dontrun{
download_session_zip() # Downloads Zip Archive from volume 31, session 9803
}
}
}
