% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_video.R
\name{download_video}
\alias{download_video}
\title{Download a specific video file.}
\usage{
download_video(asset = 1, slot = 9807, file.name = "test.mp4",
  return.response = FALSE, vb = FALSE)
}
\arguments{
\item{asset}{Asset number.}

\item{slot}{Slot/session number.}

\item{file.name}{Name for downloaded file.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}

\item{return.respose}{A Boolean value.}
}
\description{
Download a specific video file.
}
\examples{
download_video()
}
