% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_video.R
\name{download_video}
\alias{download_video}
\title{Download a specific video file.}
\usage{
download_video(vol.id = 1, session.id = 9807, asset.id = 1,
  segment.id = "-", out.dir = ".", file.name = "test.mp4",
  return.response = FALSE, vb = FALSE)
}
\arguments{
\item{vol.id}{Volume ID.}

\item{session.id}{Slot/session number.}

\item{asset.id}{Asset number.}

\item{segment.id}{Video segment.}

\item{out.dir}{Directory to save output file.}

\item{file.name}{Name for downloaded file.}

\item{return.response}{A Boolean value.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\description{
Download a specific video file.
}
\examples{
download_video()
}
