% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_video.R
\name{download_video}
\alias{download_video}
\title{Download a specific video file.}
\usage{
download_video(
  session_id = 9807,
  asset_id = 1,
  segment_id = "-",
  out_dir = ".",
  file_name = "test.mp4",
  return_response = FALSE,
  vb = FALSE
)
}
\arguments{
\item{session_id}{Slot/session number.}

\item{asset_id}{Asset number.}

\item{segment_id}{Video segment.}

\item{out_dir}{Directory to save output file.}

\item{file_name}{Name for downloaded file, default is 'test.mp4'.}

\item{return_response}{A Boolean value.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\description{
Download a specific video file.
}
\examples{
download_video()
}
