% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_volume_zip.R
\name{download_volume_zip}
\alias{download_volume_zip}
\title{Download a zip archive with all data in a volume.}
\usage{
download_volume_zip(
  vol_id = 31,
  out_dir = tempdir(),
  file_name = "test.zip",
  vb = FALSE
)
}
\arguments{
\item{vol_id}{Volume number.}

\item{out_dir}{Directory to save output file.}

\item{file_name}{Name for downloaded file, default is 'test.mp4'.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\description{
Download a zip archive with all data in a volume.
}
\examples{
download_volume_zip()
}
