% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_volume_zip.R
\name{download_volume_zip}
\alias{download_volume_zip}
\title{Download Zip Archive of All Data in a Volume.}
\usage{
download_volume_zip(
  vol_id = 31,
  out_dir = tempdir(),
  file_name = "test.zip",
  vb = FALSE
)
}
\arguments{
\item{vol_id}{Volume number.}

\item{out_dir}{Directory to save output file.}

\item{file_name}{Name for downloaded file, default is 'test.mp4'.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
Full filename of the downloaded file.
}
\description{
Download Zip Archive of All Data in a Volume.
}
\examples{
\donttest{
\dontrun{
download_volume_zip() # Zip file of all data from volume 31, the default.
}
}

}
