% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_to_csv.R
\name{dv_to_csv}
\alias{dv_to_csv}
\title{Converts a Datavyu (.opf) file to CSV.}
\usage{
dv_to_csv(dv.dir, dv.fn = "db", out.fn = paste0(dv.dir, ".csv"),
  auto.write.over = FALSE, vb = FALSE)
}
\arguments{
\item{dv.dir}{Directory to extracted Datavyu file}

\item{dv.fn}{Datavyu code file. Defaults to 'db'.}

\item{out.fn}{Output file name. Default is based on dv.dir.}

\item{auto.write.over}{A Boolean value. If TRUE, new output file overwrites old.}

\item{vb}{A boolean value. If TRUE, provides verbose output.}
}
\description{
Converts a Datavyu (.opf) file to CSV.
}
\examples{
dv_to_csv()
}
