% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_to_csv.R
\name{dv_to_csv}
\alias{dv_to_csv}
\title{Converts a Datavyu (.opf) file to CSV.}
\usage{
dv_to_csv(dv.fn, out.fn = "out.csv", auto.write.over = FALSE)
}
\arguments{
\item{dv.fn}{Full Datavyu file name.}

\item{out.fn}{Full output file name.}

\item{auto.write.over}{A Boolean value. If TRUE, new output file overwrites old.}
}
\description{
Converts a Datavyu (.opf) file to CSV.
}
\examples{
dv_to_csv()
}
