% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dv_code_defs.R
\name{extract_dv_code_defs}
\alias{extract_dv_code_defs}
\title{Converts a Datavyu (.opf) file to CSV.}
\usage{
extract_dv_code_defs(dv.fn, out.fn = "codes.csv",
  auto.write.over = TRUE)
}
\arguments{
\item{dv.fn}{Full Datavyu file name.}

\item{out.fn}{Full output file name.}

\item{auto.write.over}{A Boolean value. If TRUE, new output file overwrites old.}
}
\description{
Converts a Datavyu (.opf) file to CSV.
}
\examples{
dv_to_csv()
}
