% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_to_filetypes.R,
%   R/list_assets_in_session.R
\name{format_to_filetypes}
\alias{format_to_filetypes}
\title{Converts volume asset data frame to one with media types.}
\usage{
format_to_filetypes(vol_assets, vb = FALSE)

format_to_filetypes(vol_assets, vb = FALSE)
}
\arguments{
\item{vb}{A boolean value. If TRUE provides verbose output.}

\item{volume}{Selected volume number. Default is 2.}
}
\value{
A data framw with information about the volume owner(s).
}
\description{
Converts volume asset data frame to one with media types.

Converts volume asset data frame to one with media types.
}
\examples{
format_to_filetypes()
}
