% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_assets_in_session.R
\name{format_to_filetypes}
\alias{format_to_filetypes}
\title{Converts volume asset data frame to one with media types.}
\usage{
format_to_filetypes(vol_assets, vb = FALSE)
}
\arguments{
\item{vol_assets}{Data frame}

\item{vb}{A boolean value. If TRUE provides verbose output.}
}
\value{
A data framw with more useful information about media types.
}
\description{
Converts volume asset data frame to one with media types.
}
\examples{
format_to_filetypes ()
}
