% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asset_segment_range.R
\name{get_asset_segment_range}
\alias{get_asset_segment_range}
\title{Get the time range from a given video or audio asset.}
\usage{
get_asset_segment_range(vol_id = 1, session_id = 9807, asset_id = 1,
  vb = FALSE)
}
\arguments{
\item{vol_id}{Volume ID}

\item{session_id}{Slot/session number.}

\item{asset_id}{Asset number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\description{
Get the time range from a given video or audio asset.
}
\examples{
get_asset_segment_range()
}
