% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_asset_segment_range}
\alias{get_asset_segment_range}
\title{Get Time Range For An Asset.}
\usage{
get_asset_segment_range(
  vol_id = 1,
  session_id = 9807,
  asset_id = 1,
  convert_JSON = TRUE,
  segment_only = TRUE,
  vb = FALSE
)
}
\arguments{
\item{vol_id}{Volume ID}

\item{session_id}{Slot/session number.}

\item{asset_id}{Asset number.}

\item{convert_JSON}{A Boolean value. If TRUE, convert JSON to a data frame. Default is TRUE.}

\item{segment_only}{A Boolean value. If TRUE, returns only the segment values. Otherwise returns
a data frame with two fields, segment and permission. Default is TRUE.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
The time range (in ms) for an asset, if one is indicated.
}
\description{
Get Time Range For An Asset.
}
\examples{
\dontrun{
get_asset_segment_range()
}
}
