% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_assets.R
\name{get_assets_from_session}
\alias{get_assets_from_session}
\title{Helper function for list_volume_assets}
\usage{
get_assets_from_session(volume_container, ignore_materials = TRUE)
}
\description{
Helper function for list_volume_assets
}
