% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_stats.R
\name{get_db_stats}
\alias{get_db_stats}
\title{Get summary information about Databrary's data, institutions, or people.}
\usage{
get_db_stats(type = "stats", vb = FALSE)
}
\arguments{
\item{type}{Type of Databrary report to run "institutions", "people", "data"}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A tibble (data.frame) with the requested data.
}
\description{
Get summary information about Databrary's data, institutions, or people.
}
\examples{
get_db_stats()
get_db_stats("stats")
get_db_stats("people") # Information about the newest authorized investigators.
get_db_stats("places") # Information about the newest institutions.
}
