% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_stats.R
\name{get_db_stats}
\alias{get_db_stats}
\title{Get Stats About Databrary.}
\usage{
get_db_stats(type = "stats", vb = FALSE, rq = NULL)
}
\arguments{
\item{type}{Type of Databrary report to run "institutions", "people", "data"}

\item{vb}{A Boolean value. If TRUE provides verbose output.}

\item{rq}{An \code{httr2} request object.}
}
\value{
A data frame with the requested data or NULL if there is no new information.
}
\description{
\code{get_db_stats} returns basic summary information about
the institutions, people, and data hosted on 'Databrary.org'.
}
\examples{
\donttest{
get_db_stats()
get_db_stats("stats")
get_db_stats("people") # Information about the newest authorized investigators.
get_db_stats("places") # Information about the newest institutions.
}
}
