% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_duration.R
\name{get_file_duration}
\alias{get_file_duration}
\title{Get the the duration (in ms) of a file.}
\usage{
get_file_duration(asset_id = 1, types_w_durations = c("-600", "-800"),
  vb = FALSE)
}
\arguments{
\item{asset_id}{Asset number.}

\item{types_w_durations}{Asset types that have valid durations.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\description{
Get the the duration (in ms) of a file.
}
\examples{
get_file_duration()
}
