% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_duration}
\alias{get_file_duration}
\title{Get Duration (In ms) Of A File.}
\usage{
get_file_duration(
  asset_id = 1,
  types_w_durations = c("-600", "-800"),
  vb = FALSE
)
}
\arguments{
\item{asset_id}{Asset number.}

\item{types_w_durations}{Asset types that have valid durations.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
Duration of a file in ms.
}
\description{
Get Duration (In ms) Of A File.
}
\examples{
\dontrun{
get_file_duration() # default is the test video from databrary.org/volume/1
}
}
