% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_folder_file.R
\name{get_folder_file}
\alias{get_folder_file}
\title{Get Session File Data From A Databrary Volume}
\usage{
get_folder_file(
  vol_id = 1,
  folder_id = 9807,
  file_id,
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{vol_id}{An integer indicating the volume identifier. Default is 1.}

\item{folder_id}{An integer indicating a valid folder identifier
linked to a volume. Default value is 9807, the materials folder for volume 1.}

\item{file_id}{An integer indicating the file identifier.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{rq}{An httr2 request object.}
}
\value{
A JSON blob with the file data. If the user has previously logged
in to Databrary via \code{login_db()}, then files that have restricted access
can be downloaded, subject to the sharing release levels on those files.
}
\description{
Get Session File Data From A Databrary Volume
}
\examples{
\donttest{
\dontrun{
get_folder_file(vol_id = 2, folder_id = 11, file_id = 1)
}
}
}
