% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_sessions.R
\name{get_info_from_session}
\alias{get_info_from_session}
\title{List Sessions Info in Databrary Volume Container}
\usage{
get_info_from_session(
  volume_container,
  ignore_materials = FALSE,
  release_levels
)
}
\arguments{
\item{volume_container}{A component of a volume list returned by get_volume_by_id.}

\item{ignore_materials}{A logical value specifying whether to ignore "materials" folders.
Default is TRUE}

\item{release_levels}{A data frame mapping release level indices to release level text values.}
}
\description{
List Sessions Info in Databrary Volume Container
}
