% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_institution_by_id.R
\name{get_institution_by_id}
\alias{get_institution_by_id}
\title{Get institution metadata}
\usage{
get_institution_by_id(institution_id = 12, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{institution_id}{Institution identifier. Must be a positive integer.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}
}
\value{
List of institution metadata or NULL when inaccessible.
}
\description{
Get institution metadata
}
