% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_permission_levels}
\alias{get_permission_levels}
\title{Extract Databrary Permission Levels.}
\usage{
get_permission_levels(vb = FALSE)
}
\arguments{
\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
An array with the permission levels that can be assigned to data.
}
\description{
Extract Databrary Permission Levels.
}
\examples{
get_permission_levels()
}
