% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_permission_levels}
\alias{get_permission_levels}
\title{Extract Databrary Permission Levels.}
\usage{
get_permission_levels(vb = options::opt("vb"))
}
\arguments{
\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}
}
\value{
An array with the permission levels that can be assigned to data.
}
\description{
Extract Databrary Permission Levels.
}
\examples{
\donttest{
get_permission_levels()
}

}
