% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_release_levels.R
\name{get_release_levels}
\alias{get_release_levels}
\title{Shows release level codes}
\usage{
get_release_levels(vb = FALSE)
}
\arguments{
\item{vb}{A boolean value. If TRUE provides verbose output.}
}
\value{
Data frame with release levels.
}
\description{
Shows release level codes
}
\examples{
get_release_levels())
}
