% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_release_levels}
\alias{get_release_levels}
\title{Show Databrary Release Levels}
\usage{
get_release_levels(vb = options::opt("vb"))
}
\arguments{
\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with Databrary's release levels.
}
\description{
Show Databrary Release Levels
}
\examples{
\donttest{
get_release_levels()
}

}
