% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_file.R
\name{get_session_file}
\alias{get_session_file}
\title{Get Session File Data From A Databrary Volume}
\usage{
get_session_file(
  vol_id = 1,
  session_id = 9578,
  file_id = 27227,
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{vol_id}{An integer indicating the volume identifier. Default is 1.}

\item{session_id}{An integer indicating a valid session/slot identifier
linked to a volume. Default value is 9578.}

\item{file_id}{An integer indicating the file identifier. The default is
27227.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An httr2 request object.}
}
\value{
Metadata about the file if the user has read privileges.
}
\description{
Get Session File Data From A Databrary Volume
}
\examples{
\donttest{
\dontrun{
get_session_file(vol_id = 2, session_id = 11, file_id = 3)
# A video from volume 1, session 11.
}
}
}
