% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_supported_file_types.R
\name{get_supported_file_types}
\alias{get_supported_file_types}
\title{Extracts Databrary supported file types.}
\usage{
get_supported_file_types(vb = FALSE)
}
\arguments{
\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with the file types.
}
\description{
Extracts Databrary supported file types.
}
\examples{
get_supported_file_types()
}
