% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_supported_file_types}
\alias{get_supported_file_types}
\title{Extracts File Types Supported by Databrary.}
\usage{
get_supported_file_types(vb = FALSE)
}
\arguments{
\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with the file types permitted on Databrary.
}
\description{
Extracts File Types Supported by Databrary.
}
\examples{
\dontrun{
get_supported_file_types()
}
}
