% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_by_id.R
\name{get_user_by_id}
\alias{get_user_by_id}
\title{Get public profile information for a Databrary user}
\usage{
get_user_by_id(user_id = 6, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{user_id}{User identifier. Must be a positive integer. Default is 6.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}
}
\value{
A list with the user's public metadata.
}
\description{
Get public profile information for a Databrary user
}
