% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_video_stats.R
\name{get_video_stats}
\alias{get_video_stats}
\title{Reports summary statistics about the videos available from a given volume.}
\usage{
get_video_stats(vol_id = 1, vb = FALSE)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{vb}{A boolean value.}
}
\value{
Data frame with summary statistics about the videos in the volume.
}
\description{
Reports summary statistics about the videos available from a given volume.
}
\examples{
get_video_stats()
}
