% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_video_stats.R
\name{get_video_stats}
\alias{get_video_stats}
\title{Report Stats About Videos in a Volume.}
\usage{
get_video_stats(this_vol_id = 2, vb = FALSE)
}
\arguments{
\item{this_vol_id}{Target volume number.}

\item{vb}{A boolean value.}
}
\value{
A data frame with summary statistics about the videos in the volume.
}
\description{
Report Stats About Videos in a Volume.
}
\examples{
get_video_stats() # Retrieves summary statistics about videos in volume 2.
}
