% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_volume_by_id.R
\name{get_volume_by_id}
\alias{get_volume_by_id}
\title{Get Summary Data About A Databrary Volume}
\usage{
get_volume_by_id(vol_id = 1, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{vol_id}{Volume ID. Must be a positive integer. Default is 1.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An \code{httr2} request object. If NULL (the default), a new request
is generated using \code{make_default_request()}. To access restricted data,
the user must login with a specific request object using \code{login_db()}.}
}
\value{
A tibble with summary information about a volume.
}
\description{
Get Summary Data About A Databrary Volume
}
\examples{
\dontrun{
get_volume_by_id() # Default is Volume 1
}

}
