% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_volume_by_id.R
\name{get_volume_by_id}
\alias{get_volume_by_id}
\title{Get Data From A Databrary Volume}
\usage{
get_volume_by_id(vol_id = 1, vb = FALSE, rq = NULL)
}
\arguments{
\item{vol_id}{Volume ID.}

\item{vb}{A logical value. Show verbose messages.}

\item{rq}{An \code{httr2} request object. If NULL (the default), a new request
is generated using \code{make_default_request()}. To access restricted data,
the user must login with a specific request object using \code{login_db()}.}
}
\value{
A JSON blob with the volume data. If the user has previously logged
in to Databrary via \code{login_db()}, then volume(s) that have restricted access
can be downloaded, subject to the sharing release levels on those volume(s).
}
\description{
Get Data From A Databrary Volume
}
\examples{
\donttest{
get_volume_by_id() # Default is Volume 1
}

}
