% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HHMMSSmmm_to_ms}
\alias{HHMMSSmmm_to_ms}
\title{Get the time range from a given video or audio asset.}
\usage{
HHMMSSmmm_to_ms(HHMMSSmmm = "01:01:01:333")
}
\arguments{
\item{HHMMSSmmm}{a string in the format "HH:MM:SS:mmm"}
}
\description{
Get the time range from a given video or audio asset.
}
\examples{
HHMMSSmmm_to_ms()
}
