% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_institution.R
\name{is_institution}
\alias{is_institution}
\title{Tests whether given party ID is from an institution.}
\usage{
is_institution(party_id = 8, vb = FALSE)
}
\arguments{
\item{party_id}{Databrary party ID}

\item{vb}{A boolean value. If TRUE provides verbose output.}
}
\value{
TRUE or FALSEl.
}
\description{
Tests whether given party ID is from an institution.
}
\examples{
is_institution()
}
