% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_institution}
\alias{is_institution}
\title{Is This Party An Institution?}
\usage{
is_institution(party_id = 8, vb = FALSE, rq = NULL)
}
\arguments{
\item{party_id}{Databrary party ID}

\item{vb}{A Boolean value. If TRUE provides verbose output.}

\item{rq}{An \code{httr2} request object.}
}
\value{
TRUE if the party is an institution, FALSE otherwise.
}
\description{
Is This Party An Institution?
}
\examples{
\donttest{
is_institution() # Is party 8 (NYU) an institution.
}

}
