% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_institution.R
\name{is_person}
\alias{is_person}
\title{Tests whether given party ID is from a person.}
\usage{
is_person(party = 7, vb = vb)
}
\arguments{
\item{party}{Databrary party ID}

\item{vb}{A boolean value. If TRUE provides verbose output.}
}
\value{
TRUE or FALSEl.
}
\description{
Tests whether given party ID is from a person.
}
\examples{
is_person()
}
