% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_person}
\alias{is_person}
\title{Is This Party A Person?}
\usage{
is_person(party_id = 7, vb = FALSE)
}
\arguments{
\item{party_id}{Databrary party ID}

\item{vb}{A boolean value. If TRUE provides verbose output.}
}
\value{
TRUE or FALSE.
}
\description{
Is This Party A Person?
}
\examples{
is_person()
}
