% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_person}
\alias{is_person}
\title{Is This Party A Person?}
\usage{
is_person(party_id = 7, vb = FALSE, rq = NULL)
}
\arguments{
\item{party_id}{Databrary party ID}

\item{vb}{A boolean value. If TRUE provides verbose output.}

\item{rq}{An \code{httr2} request object.}
}
\value{
TRUE if the party is a person, FALSE otherwise.
}
\description{
Is This Party A Person?
}
\examples{
\donttest{
is_person()
}

}
