% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_assets_by_type.R
\name{list_assets_by_type}
\alias{list_assets_by_type}
\title{Lists stored assets (files) by type.}
\usage{
list_assets_by_type(vol.id = 1, type = "video", vb = FALSE)
}
\arguments{
\item{vol.id}{Volume ID}

\item{type}{Data file type, e.g. "video", "pdf", or "csv"}

\item{vb}{A boolean value. If TRUE provides verbose output.}
}
\value{
Data frame with the selected asset types.
}
\description{
Lists stored assets (files) by type.
}
\examples{
list_assets_by_type()
}
