% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_assets_by_type.R
\name{list_assets_by_type}
\alias{list_assets_by_type}
\title{List Stored Assets (Files) By Type.}
\usage{
list_assets_by_type(vol_id = 1, type = "video", vb = FALSE)
}
\arguments{
\item{vol_id}{Volume ID}

\item{type}{Data file type, e.g. "video", "pdf", or "csv"}

\item{vb}{A boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with information about assets in a volume, sorted by
type.
}
\description{
List Stored Assets (Files) By Type.
}
\examples{
list_assets_by_type() # Lists assets in volume 1.
}
