% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_assets_in_session.R
\name{list_assets_in_session}
\alias{list_assets_in_session}
\title{Assets In Databrary Session.}
\usage{
list_assets_in_session(session_id = 9807, vb = FALSE)
}
\arguments{
\item{session_id}{Slot/session ID.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with information about a selected session.
}
\description{
Assets In Databrary Session.
}
\examples{
\donttest{
list_assets_in_session() # Assets in session 9807 (from Volume 1)
#' }
}
