% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_assets_in_session.R
\name{list_assets_in_session}
\alias{list_assets_in_session}
\title{Lists assets in a given Databrary volume and session (slot).}
\usage{
list_assets_in_session(session_id = 9807, vb = FALSE)
}
\arguments{
\item{session_id}{Slot/session ID.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with the assets in the selected volume and session.
}
\description{
Lists assets in a given Databrary volume and session (slot).
}
\examples{
list_assets_in_session() # Assets in volume 1, session 9807
}
