% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_assets_in_volume.R
\name{list_assets_in_volume}
\alias{list_assets_in_volume}
\title{Assets in Databrary Volume.}
\usage{
list_assets_in_volume(vol_id = 2, vb = FALSE)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{vb}{A boolean value.}
}
\value{
A data frame with information about all assets in a volume.
}
\description{
Assets in Databrary Volume.
}
\examples{
list_assets_in_volume() # Assets in volume 2
}
