% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_assets_in_volume.R
\name{list_assets_in_volume}
\alias{list_assets_in_volume}
\title{Lists all the assets/files in a given volume.}
\usage{
list_assets_in_volume(vol.id = 1, vb = FALSE)
}
\arguments{
\item{vol.id}{Target volume number.}

\item{vb}{A boolean value.}
}
\value{
List of assets.
}
\description{
Lists all the assets/files in a given volume.
}
\examples{
list_assets_in_volume()
}
