% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_assets_json.R
\name{list_assets_json}
\alias{list_assets_json}
\title{Lists assets in a given Databrary volume and session (slot).}
\usage{
list_assets_json(session_id = 9825, vol_id = 75, vb = FALSE)
}
\arguments{
\item{session_id}{Slot/session ID.}

\item{vol_id}{Selected volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with the assets in the selected volume and session.
}
\description{
Lists assets in a given Databrary volume and session (slot).
}
\examples{
list_assets()
}
