% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_authorized_investigators.R
\name{list_authorized_investigators}
\alias{list_authorized_investigators}
\title{List authorized investigators for an institution}
\usage{
list_authorized_investigators(
  institution_id = 12,
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{institution_id}{Institution identifier.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}
}
\value{
Tibble of investigators; NULL if none.
}
\description{
List authorized investigators for an institution
}
