% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_authorized_investigators.R
\name{list_authorized_investigators}
\alias{list_authorized_investigators}
\title{List Authorized Investigators at Institution}
\usage{
list_authorized_investigators(
  party_id = 12,
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{party_id}{Target party ID.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{rq}{An \code{httr2}-style request object. If NULL, then a new request will
be generated using \code{make_default_request()}.}
}
\value{
A data frame with information the institution's authorized
investigators.
}
\description{
List Authorized Investigators at Institution
}
\examples{
\donttest{
\dontrun{
list_institutional_affiliates() # Default is Penn State (party 12)
}
}
}
