% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_individual_sponsors.R
\name{list_individual_sponsors}
\alias{list_individual_sponsors}
\title{List A Party's Individual Sponsors (people).}
\usage{
list_individual_sponsors(party_id = 406, vb = FALSE)
}
\arguments{
\item{party_id}{Target volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A tibble (data.frame) with the requested data.
}
\description{
List A Party's Individual Sponsors (people).
}
\examples{
list_individual_sponsors() # Default is Kasey Soska (party 406)
}
