% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_individual_sponsors.R
\name{list_individual_sponsors}
\alias{list_individual_sponsors}
\title{List A Party's Individual Sponsors.}
\usage{
list_individual_sponsors(
  party_id = 406,
  report_target_party = TRUE,
  vb = FALSE
)
}
\arguments{
\item{party_id}{Party number. Default is 406 (Kasey Soska)}

\item{report_target_party}{Print the party ID and name of the target person.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A tibble (data.frame) with the requested data.
}
\description{
A party (person or institution) may have a sponsor that is another person
or an institution. This function lists the \emph{person(s)} who sponsor
a party.
}
\examples{
list_individual_sponsors() # Default is Kasey Soska (party 406)
}
