% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_institution_affiliates.R
\name{list_institution_affiliates}
\alias{list_institution_affiliates}
\title{List affiliates for an institution}
\usage{
list_institution_affiliates(
  institution_id = 12,
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{institution_id}{Institution identifier.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}
}
\value{
Tibble of affiliates with roles and expiration dates.
}
\description{
List affiliates for an institution
}
