% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_institutional_sponsors.R
\name{list_institutional_sponsors}
\alias{list_institutional_sponsors}
\title{Lists a party's institutional sponsors.}
\usage{
list_institutional_sponsors(party_id = 6, vb = FALSE)
}
\arguments{
\item{party_id}{Target volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A tibble (data.frame) with the requested data.
}
\description{
Lists a party's institutional sponsors.
}
\examples{
list_institutional_sponsors() # Defaults to Rick Gilmore (party 6)
}
