% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_institutional_sponsors.R
\name{list_institutional_sponsors}
\alias{list_institutional_sponsors}
\title{List Institutional Sponsors For A Party.}
\usage{
list_institutional_sponsors(
  party_id = 6,
  report_target_party = TRUE,
  vb = FALSE
)
}
\arguments{
\item{party_id}{Target volume number.}

\item{report_target_party}{A Boolean value. Print info about the target party.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with information about a party's \emph{institutional} sponsor(s).
}
\description{
A party (person or institution) may have a sponsor that is another person
or an institution. This function lists the \emph{institution(s)} that sponsor
a party.
}
\examples{
\donttest{
list_institutional_sponsors() # Defaults to Rick Gilmore (party 6)
}
}
