% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_party.R
\name{list_party}
\alias{list_party}
\title{List Data For A Party (person or institution).}
\usage{
list_party(party_id = 8, component = "all", vb = FALSE)
}
\arguments{
\item{party_id}{Target volume number. Default is volume 8 (NYU).}

\item{component}{Which data to return 'children', 'parents', or 'all'. Default is all.}

\item{vb}{A Boolean value. If TRUE provides verbose output. Default is FALSE.}
}
\value{
A data frame with information about a party (person or institution).
}
\description{
List Data For A Party (person or institution).
}
\examples{
list_party() # Default is New York University (party 8)
}
