% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_people.R
\name{list_people}
\alias{list_people}
\title{Lists basic information about people on Databrary.}
\usage{
list_people(people_list = 5:7, vb = FALSE)
}
\arguments{
\item{people_list}{Party number(s) for people to list.}

\item{vb}{A Boolean value if TRUE returns verbose output.}
}
\value{
A data frame with the people in the specified range of party ID's.
}
\description{
Lists basic information about people on Databrary.
}
\examples{
list_people() # Lists people with party IDs 5:7, Databrary's founders
}
