% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_people.R
\name{list_people}
\alias{list_people}
\title{Lists basic information about people on Databrary.}
\usage{
list_people(people.list = 7)
}
\arguments{
\item{people.list}{Party number(s) for people to list.}
}
\value{
Status code if successful.
}
\description{
Lists basic information about people on Databrary.
}
\examples{
list_people()
}
