% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_session_activity.R
\name{list_session_activity}
\alias{list_session_activity}
\title{List activity in a given Databrary session for which the user has
ownership privileges.}
\usage{
list_session_activity(session_id = 6256, vb = FALSE)
}
\arguments{
\item{session_id}{Selected session/slot number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}

\item{vol_id}{Selected volume number.}
}
\value{
A list with the activity history on a slot.
}
\description{
List activity in a given Databrary session for which the user has
ownership privileges.
}
\examples{
list_session_activity()
}
