% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_session_activity.R
\name{list_session_activity}
\alias{list_session_activity}
\title{List Activity History in Databrary Session.}
\usage{
list_session_activity(session_id = 6256, vb = FALSE)
}
\arguments{
\item{session_id}{Selected session/slot number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with the activity history on a session/slot.
}
\description{
If a user has access to a volume and session, this function returns the
history of modifications to that session.
}
\examples{
\donttest{
\dontrun{
list_session_activity(session_id = 6256) # History of session 6256. Only
shown to researchers who have write privileges on this session.
}
}
}
