% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_session_assets.R
\name{list_session_assets}
\alias{list_session_assets}
\title{List Assets in a Databrary Session.}
\usage{
list_session_assets(
  session_id = 9807,
  vol_id = NULL,
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{session_id}{An integer. A Databrary session number. Default is 9807,
the "materials" folder from Databrary volume 1.}

\item{vol_id}{Optional integer. The volume containing the session. Recent
versions of the Databrary API require this value to be supplied because
session identifiers are scoped to volumes.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An \code{httr2} request object. If NULL, a default request is generated
from databraryr::make_default_request().}
}
\value{
A data frame with information about all assets in a volume.
}
\description{
#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
As of v0.6.3 \code{list_session_assets()} replaces an older function that is now
named \code{list_volume_session_assets()}.  The older function requires both
a volume ID and a session ID. The new function requires only a session ID.
}
\examples{
\donttest{
\dontrun{
list_session_assets() # Session 9807 in volume 1
}
}
}
