% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_session_assets.R
\name{list_session_assets}
\alias{list_session_assets}
\title{List Assets in a Session from a Databrary volume.}
\usage{
list_session_assets(vol_id = 1, session_id = 9807, vb = FALSE, rq = NULL)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{session_id}{The session number in the selected volume.}

\item{vb}{A boolean value.}

\item{rq}{An \code{httr2} request object.}
}
\value{
A data frame with information about all assets in a volume.
}
\description{
List Assets in a Session from a Databrary volume.
}
\examples{
\donttest{
\dontrun{
list_session_assets() # Session 9807 in volume 1
}
}
}
