% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sessions.R
\name{list_sessions}
\alias{list_sessions}
\title{Download Information About A Volume or Volume's Sessions.}
\usage{
list_sessions(vol_id = 1, vb = FALSE)
}
\arguments{
\item{vol_id}{Target volume ID. Defaults to 1.}

\item{vb}{A boolean value.}
}
\value{
Information about sessions in a volume.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{list_sessions()} has been deprecated in favor of \code{list_volume_sessions()}.
}
\examples{
\donttest{
list_sessions() # Lists sessions in volume 1.
}
}
