% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sessions.R
\name{list_sessions}
\alias{list_sessions}
\title{Downloads session spreadsheet as a CSV.}
\usage{
list_sessions(volume = 1, vb = vb)
}
\arguments{
\item{volume}{Target volume number.}

\item{vb}{A boolean value.}
}
\value{
List of assets.
}
\description{
Downloads session spreadsheet as a CSV.
}
\examples{
list_sessions()
}
