% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sessions.R
\name{list_sessions}
\alias{list_sessions}
\title{Downloads list of sessions.}
\usage{
list_sessions(vol.id = 1, vb = FALSE)
}
\arguments{
\item{vol.id}{Target volume ID. Defaults to 1.}

\item{vb}{A boolean value.}
}
\value{
List of sessions.
}
\description{
Downloads list of sessions.
}
\examples{
list_sessions()
}
