% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sessions_in_volume.R
\name{list_sessions_in_volume}
\alias{list_sessions_in_volume}
\title{Lists the sessions in a given volume. This is an alias for list_sessions().}
\usage{
list_sessions_in_volume(vol_id = 1, vb = FALSE)
}
\arguments{
\item{vol_id}{Target volume ID. Defaults to 1.}

\item{vb}{A boolean value.}
}
\value{
List of sessions.
}
\description{
Lists the sessions in a given volume. This is an alias for list_sessions().
}
\examples{
list_sessions_in_volume() # Lists sessions in volume 1.
}
