% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sponsors.R
\name{list_sponsors}
\alias{list_sponsors}
\title{List Sponsors For A Party}
\usage{
list_sponsors(party_id = 6, vb = FALSE)
}
\arguments{
\item{party_id}{Target party ID.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A tibble (data.frame) with information about a party's sponsors.
}
\description{
List Sponsors For A Party
}
\examples{
\dontrun{
list_sponsors() # Default is Rick Gilmore (party 6)
}
}
