% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_user_affiliates.R
\name{list_user_affiliates}
\alias{list_user_affiliates}
\title{List affiliates for a user}
\usage{
list_user_affiliates(user_id = 6, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{user_id}{User identifier.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}
}
\value{
Tibble of affiliates for the user.
}
\description{
List affiliates for a user
}
