% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_user_volumes.R
\name{list_user_volumes}
\alias{list_user_volumes}
\title{List volumes associated with a user}
\usage{
list_user_volumes(user_id = 6, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{user_id}{User identifier.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}
}
\value{
Tibble of volumes the user owns or collaborates on.
}
\description{
List volumes associated with a user
}
