% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume.R
\name{list_volume}
\alias{list_volume}
\title{List Info for a Databrary Volume.}
\usage{
list_volume(vol_id = 1, search_str = "", vb = FALSE)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{search_str}{The specific data to retrieve for the volume}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A list (or data frame) with the requested data.
}
\description{
List Info for a Databrary Volume.
}
\examples{
list_volume() # Default is volume 1
}
