% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_activity.R
\name{list_volume_activity}
\alias{list_volume_activity}
\title{List activity in a given Databrary volume.}
\usage{
list_volume_activity(vol_id = 1, vb = FALSE)
}
\arguments{
\item{vol_id}{Selected volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A list with the activity history on a volume.
}
\description{
List activity in a given Databrary volume.
}
\examples{
list_volume_activity()
}
