% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_activity.R
\name{list_volume_activity}
\alias{list_volume_activity}
\title{List Activity In A Databrary Volume}
\usage{
list_volume_activity(vol_id = 1, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{vol_id}{Selected volume number.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{rq}{An \code{httr2} request object. Defaults to NULL.}
}
\value{
A list with the activity history on a volume.
}
\description{
If a user has access to a volume, this command lists the modification
history of the volume as a
}
\examples{
\donttest{
\dontrun{
# The following will only return output if the user has write privileges
# on the volume.

list_volume_activity(vol_id = 1) # Activity on volume 1.
}
}
}
