% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_activity.R
\name{list_volume_activity}
\alias{list_volume_activity}
\title{List Activity In A Databrary Volume}
\usage{
list_volume_activity(vol_id = 1, vb = FALSE)
}
\arguments{
\item{vol_id}{Selected volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A list with the activity history on a volume.
}
\description{
If a user has access to a volume, this command lists the modification
history of the volume.
}
\examples{
\dontrun{
list_volume_activity() # Lists activity on Volume 1.
}
}
