% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_assets.R
\name{list_volume_assets}
\alias{list_volume_assets}
\title{List Assets in Databrary Volume.}
\usage{
list_volume_assets(vol_id = 1, vb = FALSE, rq = NULL)
}
\arguments{
\item{vol_id}{Target volume number. Default is 1.}

\item{vb}{A logical value. Default is FALSE.}

\item{rq}{An \code{httr2} request object. Default is NULL.}
}
\value{
A data frame with information about all assets in a volume.
}
\description{
List Assets in Databrary Volume.
}
\examples{
\donttest{
\dontrun{
list_volume_assets() # Assets in volume 1
}
}
}
