% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_excerpts.R
\name{list_volume_excerpts}
\alias{list_volume_excerpts}
\title{List Image or Video Excerpts On A Databrary Volume.}
\usage{
list_volume_excerpts(vol_id = 1, vb = FALSE)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with information about any available excerpts.
}
\description{
List Image or Video Excerpts On A Databrary Volume.
}
\examples{
\donttest{
list_volume_excerpts()
}
}
