% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_funding.R
\name{list_volume_funding}
\alias{list_volume_funding}
\title{Lists Funders Associated With a Databrary Volume.}
\usage{
list_volume_funding(vol_id = 1, vb = FALSE)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A tibble (data.frame) with the requested data.
}
\description{
Lists Funders Associated With a Databrary Volume.
}
\examples{
list_volume_funding()
}
