% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_info.R
\name{list_volume_info}
\alias{list_volume_info}
\title{List Basic Volume Info.}
\usage{
list_volume_info(vol_id = 1, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{rq}{An \code{httr2} request object. If NULL (the default)
a request will be generated, but this will only permit public information
to be returned.}
}
\value{
A data frame with basic information about a volume.
}
\description{
List Basic Volume Info.
}
\examples{
\donttest{
\dontrun{
list_volume_info() # Sessions in Volume 1
}
}

}
