% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_links.R
\name{list_volume_links}
\alias{list_volume_links}
\title{Retrieves URL links associated with a Databrary volume.}
\usage{
list_volume_links(vol_id = 1, vb = FALSE)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A tibble (data.frame) with the requested data.
}
\description{
Retrieves URL links associated with a Databrary volume.
}
