% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_metadata.R
\name{list_volume_metadata}
\alias{list_volume_metadata}
\title{List volume metadata.}
\usage{
list_volume_metadata(
  vol_id = 2,
  write_header = FALSE,
  data_frame = TRUE,
  vb = FALSE
)
}
\arguments{
\item{vol_id}{Selected volume number.}

\item{write_header}{A Boolean value. If TRUE writes a comma-separated header.}

\item{data_frame}{A Boolean value. If TRUE writes a data frame as output.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with information about the volume.
}
\description{
List volume metadata.
}
