% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_metadata.R
\name{list_volume_metadata}
\alias{list_volume_metadata}
\title{List Volume Metadata.}
\usage{
list_volume_metadata(vol_id = 1, vb = FALSE)
}
\arguments{
\item{vol_id}{Selected volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with metadata about a volume.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has been deprecated and may be removed in a future release.
See \code{get_volume_by_id()} and other \verb{list_volume_*} functions for related
functionality.
}
\examples{
\donttest{
list_volume_metadata() # List info about volume 1.
}
}
