% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_metadata.R
\name{list_volume_metadata}
\alias{list_volume_metadata}
\title{List Volume Metadata.}
\usage{
list_volume_metadata(vol_id = 1, vb = FALSE)
}
\arguments{
\item{vol_id}{Selected volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with metadata about a volume.
}
\description{
List Volume Metadata.
}
\examples{
\dontrun{
list_volume_metadata() # List info about volume 1.
}
}
