% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_owners.R
\name{list_volume_owners}
\alias{list_volume_owners}
\title{List vol.id owners.}
\usage{
list_volume_owners(this_vol_id = 1, vb = FALSE)
}
\arguments{
\item{vb}{A boolean value. If TRUE provides verbose output.}

\item{vol_id}{Selected volume number. Default is 2.}
}
\value{
A data framw with information about the volume owner(s).
}
\description{
List vol.id owners.
}
\examples{
list_volume_owners()
}
