% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_owners.R
\name{list_volume_owners}
\alias{list_volume_owners}
\title{List Owners of a Databrary Volume.}
\usage{
list_volume_owners(this_vol_id = 1, vb = FALSE)
}
\arguments{
\item{this_vol_id}{Selected volume number. Default is volume 1.}

\item{vb}{A boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with information about a volume's owner(s).
}
\description{
List Owners of a Databrary Volume.
}
\examples{
list_volume_owners()
}
