% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_session_assets.R
\name{list_volume_session_assets}
\alias{list_volume_session_assets}
\title{List Assets in a Session from a Databrary volume.}
\usage{
list_volume_session_assets(
  vol_id = 2,
  session_id = 11,
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{vol_id}{Target volume number. Must be a positive integer.}

\item{session_id}{The session number in the selected volume.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An \code{httr2} request object.}
}
\value{
A data frame with information about all assets in a volume.
}
\description{
#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
\code{list_volume_session_assets()} is a new name for the <v0.6.2 \code{list_session_assets()} as
function. There is a new \code{list_session_assets()} function that does not
requre the volume ID. The \code{list_volume_session_assets()} \emph{requires} a volume
ID.
}
\examples{
\donttest{
\dontrun{
list_volume_session_assets() # Defaults to session 11 in volume 2
}
}
}
