% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_session_assets.R
\name{list_volume_session_assets}
\alias{list_volume_session_assets}
\title{List Assets in a Session from a Databrary volume.}
\usage{
list_volume_session_assets(
  vol_id = 1,
  session_id = 9807,
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{session_id}{The session number in the selected volume.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{rq}{An \code{httr2} request object.}
}
\value{
A data frame with information about all assets in a volume.
}
\description{
#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
\code{list_volume_session_assets()} is a new name for the <v0.6.2 \code{list_session_assets()} as
function. There is a new \code{list_session_assets()} function that does not
requre the volume ID. The \code{list_volume_session_assets()} \emph{requires} a volume
ID.
}
\examples{
\donttest{
\dontrun{
list_volume_session_assets() # Session 9807 in volume 1
}
}
}
