% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_tags.R
\name{list_volume_tags}
\alias{list_volume_tags}
\title{Lists Keywords And Tags For A Volume.}
\usage{
list_volume_tags(vol_id = 1, vb = FALSE)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with the requested data.
}
\description{
Lists Keywords And Tags For A Volume.
}
\examples{
\dontrun{
list_volume_tags()
}
}
