% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_volume_tags.R
\name{list_volume_tags}
\alias{list_volume_tags}
\title{Lists Keywords And Tags For A Volume.}
\usage{
list_volume_tags(vol_id = 1, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{vol_id}{Target volume number.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{rq}{An \code{httr2} request object. Default is NULL.}
}
\value{
A data frame with the requested data.
}
\description{
Lists Keywords And Tags For A Volume.
}
\examples{
\donttest{
list_volume_tags()
}
}
