% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_db.R
\name{login_db}
\alias{login_db}
\title{Log In To Databrary.org.}
\usage{
login_db(
  email = NULL,
  password = NULL,
  client_id = NULL,
  client_secret = NULL,
  store = FALSE,
  overwrite = FALSE,
  SERVICE = KEYRING_SERVICE,
  vb = options::opt("vb")
)
}
\arguments{
\item{email}{Databrary account email address.}

\item{password}{Databrary password (not recommended as it will displayed
as you type)}

\item{client_id}{OAuth2 client identifier.}

\item{client_secret}{OAuth2 client secret.}

\item{store}{A boolean value. If TRUE store/retrieve credentials from the
system keyring/keychain.}

\item{overwrite}{A boolean value. If TRUE and store is TRUE, overwrite/
update stored credentials in keyring/keychain.}

\item{SERVICE}{A character label for stored credentials in the keyring.
Default is \code{org.databrary.databraryr}.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}
}
\value{
Logical value indicating whether log in is successful or not.
}
\description{
Log In To Databrary.org.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
login_db() # Queries user for email and password interactively.
\dontshow{\}) # examplesIf}
\donttest{
\dontrun{
# The following shows how to use credentials that have been stored previously.

login_db(email = "you@provider.com", store = TRUE)

}
}
}
