% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_db.R
\name{login_db}
\alias{login_db}
\title{Logs in to Databrary.org.}
\usage{
login_db(email = NULL, login.url = "/api/user/login",
  return.response = FALSE, save.session = TRUE,
  stored.credentials = FALSE, system.credentials = TRUE,
  credentials.file = "~/api-keys/json/databrary-keys.json", vb = FALSE)
}
\arguments{
\item{email}{Databrary account ID (email).}

\item{login.url}{URL for the login API.}

\item{save.session}{A Boolean value. If TRUE saves the session cookie.}

\item{stored.credentials}{A Boolean value. If TRUE loads login credentials from credentials file.}

\item{credentials.file}{Path to stored credentials file if it exists.}

\item{vb}{A boolean value. If TRUE provides verbose output.}

\item{return.reponse}{A Boolean value. If TRUE returns results of GET}
}
\value{
Status code if login is successful
}
\description{
Logs in to Databrary.org.
}
\examples{
login_db()
}
