% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_db.R
\name{login_db}
\alias{login_db}
\title{Logs In to Databrary.org.}
\usage{
login_db(
  email = NULL,
  password = NULL,
  store = FALSE,
  overwrite = FALSE,
  vb = FALSE
)
}
\arguments{
\item{email}{Databrary account email address.}

\item{password}{Databrary password (not recommended as it will displayed as you type)}

\item{store}{A boolean value. If TRUE store/retrieve credentials from the system keyring/keychain.}

\item{overwrite}{A boolean value. If TRUE and store is TRUE, overwrite/ update stored credentials in keyring/keychain.}

\item{vb}{A boolean value. If TRUE provides verbose output.}
}
\value{
Logical value indicating whether login is successful or not.
}
\description{
Logs In to Databrary.org.
}
\examples{
login_db()
}
