% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout_db.R
\name{logout_db}
\alias{logout_db}
\title{Log Out of Databrary.org.}
\usage{
logout_db(vb = options::opt("vb"))
}
\arguments{
\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}
}
\value{
TRUE if logging out succeeds, FALSE otherwise.
}
\description{
Log Out of Databrary.org.
}
\examples{
\donttest{
logout_db()
}
}
