% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout_db.R
\name{logout_db}
\alias{logout_db}
\title{Logs out of Databrary.org.}
\usage{
logout_db(vb = TRUE)
}
\arguments{
\item{vb}{A boolean value. If TRUE provides verbose output.}

\item{logout.url}{URL for the login API.}
}
\value{
Status code if successful.
}
\description{
Logs out of Databrary.org.
}
\examples{
logout_db()
}
