% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout_db.R
\name{logout_db}
\alias{logout_db}
\title{Log Out of Databrary.org.}
\usage{
logout_db(vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{vb}{A boolean value. If TRUE provides verbose output.}

\item{rq}{An \code{httr2} request object. Defaults to NULL.}
}
\value{
TRUE if logging out succeeds, FALSE otherwise.
}
\description{
Log Out of Databrary.org.
}
\examples{
\donttest{
logout_db()
}
}
