% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_default_request.R
\name{make_default_request}
\alias{make_default_request}
\title{Set default httr request parameters.}
\usage{
make_default_request()
}
\value{
An \code{httr2} request object
}
\description{
\code{make_default_request} sets default parameters for httr requests.
}
\examples{
make_default_request()
}
