% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_fn_portable}
\alias{make_fn_portable}
\title{Make Portable File Names}
\usage{
make_fn_portable(
  fn,
  vb = options::opt("vb"),
  replace_regex = "[ &\\\\!\\\\)\\\\(\\\\}\\\\{\\\\[\\\\]\\\\+\\\\=@#\\\\$\%\\\\^\\\\*]",
  replacement_char = "_"
)
}
\arguments{
\item{fn}{Databrary party ID}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{replace_regex}{A character string. A regular expression to capture
the "non-portable" characters in fn.}

\item{replacement_char}{A character string. The character(s) that will replace
the non-portable characters.}
}
\value{
A "cleaned" portable file name
}
\description{
Make Portable File Names
}
