% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_data_as_df.R
\name{read_csv_data_as_df}
\alias{read_csv_data_as_df}
\title{Download Stored CSV As Data Frame.}
\usage{
read_csv_data_as_df(session_id = 9807, asset_id = 153108, vb = FALSE)
}
\arguments{
\item{session_id}{Session ID within some volume.}

\item{asset_id}{Asset ID within the session.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with the CSV data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has been deprecated and may be removed in a future release.
See \code{download_session_asset()} for more similar, but more general functionality.
}
\details{
If a given Databrary session has a CSV file, this function will
try to download it as a data frame.
}
\examples{
\donttest{
read_csv_data_as_df() # Downloads a CSV depicting Databrary growth from volume 1
}
}
