% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_funder.R
\name{search_for_funder}
\alias{search_for_funder}
\title{Report Information About A Funder.}
\usage{
search_for_funder(
  search_string = "national science foundation",
  approved_only = TRUE,
  vb = options::opt("vb"),
  rq = NULL
)
}
\arguments{
\item{search_string}{String to search. Default is "national science foundation".}

\item{approved_only}{Logical. When TRUE (default) only approved funders are
returned. Set to FALSE to include unapproved funders as well.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An \code{httr2} request object. Default is NULL.}
}
\value{
A data frame with information about the funder.
}
\description{
Report Information About A Funder.
}
\examples{
\dontrun{
search_for_funder("national+science+foundation")
}

}
