% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_funder.R
\name{search_for_funder}
\alias{search_for_funder}
\title{Report Information About A Funder.}
\usage{
search_for_funder(search_string = "national+science+foundation", vb = FALSE)
}
\arguments{
\item{search_string}{String to search.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A data frame with information about the funder.
}
\description{
Report Information About A Funder.
}
\examples{
search_for_funder()
}
