% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_funder.R
\name{search_for_funder}
\alias{search_for_funder}
\title{Search for funders.}
\usage{
search_for_funder(search_string = "national+science+foundation", vb = FALSE)
}
\arguments{
\item{vb}{A Boolean value. If TRUE provides verbose output.}

\item{keyword_string}{String to search.}
}
\value{
A list with the activity history on a slot.
}
\description{
Search for funders.
}
\examples{
search_for_funder()
}
