% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_funder.R
\name{search_for_funder}
\alias{search_for_funder}
\title{Report Information About A Funder.}
\usage{
search_for_funder(
  search_string = "national+science+foundation",
  vb = FALSE,
  rq = NULL
)
}
\arguments{
\item{search_string}{String to search.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}

\item{rq}{An \code{httr2} request object. Default is NULL.}
}
\value{
A data frame with information about the funder.
}
\description{
Report Information About A Funder.
}
\examples{
\donttest{
search_for_funder("national+science+foundation")
}

}
