% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_keywords.R
\name{search_for_keywords}
\alias{search_for_keywords}
\title{Search for keywords in Databrary volumes.}
\usage{
search_for_keywords(search_string = "locomotion", vb = FALSE)
}
\arguments{
\item{search_string}{String to search.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A list with the activity history on a slot.
}
\description{
Search for keywords in Databrary volumes.
}
\examples{
search_for_keywords()
}
