% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_keywords.R
\name{search_for_keywords}
\alias{search_for_keywords}
\title{Search For Keywords in Databrary Volumes.}
\usage{
search_for_keywords(search_string = "locomotion", vb = FALSE)
}
\arguments{
\item{search_string}{String to search.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
A list with the volumes that contain the keyword.
}
\description{
Search For Keywords in Databrary Volumes.
}
\examples{
\dontrun{
search_for_keywords() # searches for volumes with "locomotion" as a keyword.
search_for_keywords("adult") # searches for volumes with "adult" as a keyword.
#' }
}
