% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_tags.R
\name{search_for_tags}
\alias{search_for_tags}
\title{Search For Tags on Volumes or Sessions.}
\usage{
search_for_tags(search_string = "ICIS", vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{search_string}{String to search.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{rq}{An \code{httr2} request object. Default is NULL.}
}
\value{
An array of tags that match the tag_string.
}
\description{
Search For Tags on Volumes or Sessions.
}
\examples{
\dontrun{
search_for_tags() # Searches for volumes that have the tag "ICIS"
}

}
