% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_tags.R
\name{search_for_tags}
\alias{search_for_tags}
\title{Search For Tags on Volumes or Sessions.}
\usage{
search_for_tags(search_string = "ICIS", vb = FALSE)
}
\arguments{
\item{search_string}{String to search.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
An array of tags that match the tag_string.
}
\description{
Search For Tags on Volumes or Sessions.
}
\examples{
search_for_tags()
}
