% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_tags.R
\name{search_for_tags}
\alias{search_for_tags}
\title{Search for tags on volumes or sessions.}
\usage{
search_for_tags(search_string = "ICIS", vb = FALSE)
}
\arguments{
\item{vb}{A Boolean value. If TRUE provides verbose output.}

\item{tag_string.}{String to search.}
}
\value{
An array of tags that match the tag_string.
}
\description{
Search for tags on volumes or sessions.
}
\examples{
search_for_tags()
}
