% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_institutions.R
\name{search_institutions}
\alias{search_institutions}
\title{Search For Institutions In Databrary.}
\usage{
search_institutions(search_string, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{search_string}{Character string describing the institution search
query.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}

\item{rq}{An \code{httr2} request object. Defaults to \code{NULL}.}
}
\value{
A tibble containing matching institutions ordered by relevance, or
\code{NULL} when no matches exist for the query.
}
\description{
Perform a search across institutions registered with
Databrary.
}
\examples{
\donttest{
\dontrun{
search_institutions("state")
}
}
}
