% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_users.R
\name{search_users}
\alias{search_users}
\title{Search For Users In Databrary.}
\usage{
search_users(search_string, vb = options::opt("vb"), rq = NULL)
}
\arguments{
\item{search_string}{Character string describing the search query.}

\item{vb}{Show verbose feedback. Defaults to \code{options::opt("vb")}.}

\item{rq}{An \code{httr2} request object. Defaults to \code{NULL}.}
}
\value{
A tibble containing user matches ordered by relevance, or \code{NULL}
when no matches exist for the query.
}
\description{
Perform a directory search across Databrary users by name or
email address.
}
\examples{
\donttest{
\dontrun{
search_users("gilmore")
}
}
}
