% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_videos_in_volume.R
\name{summarize_videos_in_volume}
\alias{summarize_videos_in_volume}
\title{Summarize Data About Videos in Databrary Volume.}
\usage{
summarize_videos_in_volume(vol_id = 2, vb = FALSE)
}
\arguments{
\item{vol_id}{An integer. Selected volume number or numbers.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}
}
\value{
Data frame with the vol_id, number of videos, and total hours.
}
\description{
Summarize Data About Videos in Databrary Volume.
}
\examples{
\donttest{
\dontrun{
summarize_videos_in_volume(vol_id = 2) # Summary of volume 2

summarize_videos_in_volume(vol_id = c(4, 5)) # Summarize volumes 4 & 5
}
}
}
