% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_volume.R
\name{summarize_volume}
\alias{summarize_volume}
\title{Plot summary of volume's participant characteristics.}
\usage{
summarize_volume(vol.id = 4, plot.style = "ggplot",
  return.df = FALSE, vb = FALSE)
}
\arguments{
\item{plot.style}{Defaults to 'ggplot'}

\item{return.df}{A Boolean value. If TRUE returns a data frame.}

\item{vb}{A Boolean value. If TRUE provides verbose output.}

\item{volume}{Selected volume number.}
}
\value{
Status code if successful.
}
\description{
Plot summary of volume's participant characteristics.
}
\examples{
summarize_volume()
}
