% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoami.R
\name{whoami}
\alias{whoami}
\title{Retrieve metadata about the authenticated Databrary user.}
\usage{
whoami(refresh = TRUE, vb = options::opt("vb"))
}
\arguments{
\item{refresh}{Whether to attempt automatic token refresh when the current
access token is expired. Defaults to \code{TRUE}.}

\item{vb}{Show verbose messages. (Defaults to \code{FALSE}, overwritable using option 'databraryr.vb' or environment variable 'R_DATABRARYR_VB')}
}
\value{
A list containing \code{auth_method} and \code{user} fields (both lists) or
\code{NULL} if the request fails due to lack of authentication.
}
\description{
Calls the Django \verb{/oauth2/test/} endpoint to report the current authentication
method and user profile. Requires a valid OAuth2 access token acquired via
\code{login_db()}.
}
\examples{
\\dontrun{
login_db()
whoami()
}
}
